zng_app::crash_handler

Struct CrashConfig

pub struct CrashConfig { /* private fields */ }

Crash handler config.

Use crash_handler_config! to set config.

Implementations

impl CrashConfig

pub fn dialog(&mut self, handler: impl FnOnce(CrashArgs) + 'static)

Set the crash dialog process handler.

The dialog handler can run an app or show a native dialog, it must use the CrashArgs process terminating methods to respond, if it returns CrashArgs::exit will run.

Note that the handler does not need to actually show any dialog, it can just save crash info and restart the app for example.

pub fn default_dialog(&mut self, handler: impl FnOnce(CrashArgs) + 'static)

Set the crash dialog-handler used if crash_dialog is not set.

This is used by app libraries or themes to provide a default dialog.

pub fn app_process( &mut self, cfg: impl for<'a, 'b> FnMut(&'a mut Command, &'b CrashArgs) -> &'a mut Command + 'static, )

Add a closure that is called just before the app-process is spawned.

pub fn dialog_process( &mut self, cfg: impl for<'a, 'b> FnMut(&'a mut Command, &'b CrashArgs) -> &'a mut Command + 'static, )

Add a closure that is called just before the dialog-process is spawned.

pub fn minidump_dir(&mut self, dir: impl Into<PathBuf>)

Change the minidump directory.

Is zng::env::cache("zng_minidump") by default.

pub fn no_minidump(&mut self)

Do not collect a minidump.

pub fn no_crash_handler(&mut self)

Does not run with crash handler.

This is equivalent of running with NO_ZNG_CRASH_HANDLER env var.