Struct zng_app::event::AnyEvent

pub struct AnyEvent { /* private fields */ }

Represents an Event without the args type.

Implementations

impl AnyEvent

pub fn name(&self) -> &'static str

Name of the event static item.

pub fn is<A: EventArgs>(&self, event: &Event<A>) -> bool

Returns true if self is the type erased event.

pub fn has(&self, update: &EventUpdate) -> bool

Returns true if the update is for this event.

pub fn hook( &self, hook: impl Fn(&mut EventUpdate) -> bool + Send + Sync + 'static ) -> EventHandle

Register a callback that is called just before an event begins notifying.

pub fn subscribe(&self, widget_id: WidgetId) -> EventHandle

Register the widget to receive targeted events from this event.

Widgets only receive events if they are in the delivery list generated by the event args and are subscribers to the event. App extensions receive all events.

pub fn is_subscriber(&self, widget_id: WidgetId) -> bool

Returns true if the widget is subscribed to this event.

pub fn has_subscribers(&self) -> bool

Returns true if at least one widget is subscribed to this event.

pub fn visit_subscribers(&self, visit: impl FnMut(WidgetId))

Calls visit for each widget subscribed to this event.

Note that trying to subscribe or add hook inside visit will deadlock. Inside visit you can notify the event and generate event updates.

pub fn has_hooks(&self) -> bool

Returns true if any app level callback is registered for this event.

This includes AnyEvent::hook, Event::on_pre_event, Event::on_event and Event::receiver.