zng_app::event
Struct Command

pub struct Command { /* private fields */ }

Identifies a command event.

Use the command! to declare commands, it declares command static items with optional metadata initialization.

command! {
    /// Foo-bar command.
    pub static FOO_BAR_CMD = {
        foo: true,
        bar: false,
    };
}

Metadata

Commands can have associated metadata, this metadata is extendable and can be used to enable command features such as command shortcuts. The metadata can be accessed using with_meta, metadata extensions traits can use this metadata to store state. See CommandMeta for more details.

Handles

Unlike other events, commands only notify if it has at least one handler, handlers must call subscribe to indicate that the command is relevant to the current app state and set the subscription handle enabled flag to indicate that the handler can fulfill command requests.

Scopes

Commands are global by default, meaning an enabled handle anywhere in the app enables it everywhere. You can use scoped to declare sub-commands that are the same command event, but filtered to a scope, metadata of scoped commands inherit from the app scope metadata, but can be overridden just for the scope.

Implementations

impl Command

pub fn subscribe(&self, enabled: bool) -> CommandHandle

Create a new handle to this command.

A handle indicates that command handlers are present in the current app, the enabled flag indicates the handler is ready to fulfill command requests.

If the command is scoped on a window or widget if it is added to the command event subscribers.

pub fn subscribe_wgt(&self, enabled: bool, target: WidgetId) -> CommandHandle

Create a new handle for this command for a handler in the target widget.

The handle behaves like subscribe, but include the target on the delivery list for app scoped commands. Note this only works for global commands (app scope), window and widget scoped commands only notify the scope so the target is ignored for scoped commands.

pub fn event(&self) -> Event<CommandArgs>

Underlying event that represents this command in any scope.

pub fn scope(&self) -> CommandScope

Command scope.

pub fn scoped(self, scope: impl Into<CommandScope>) -> Command

Gets the command in a new scope.

pub fn with_meta<R>(&self, visit: impl FnOnce(&mut CommandMeta<'_>) -> R) -> R

Visit the command custom metadata of the current scope.

Metadata for CommandScope::App is retained for the duration of the app, metadata scoped on window or widgets is dropped after an update cycle with no handler and no strong references to has_handlers and is_enabled.

pub fn has(&self, update: &EventUpdate) -> bool

Returns true if the update is for this command and scope.

pub fn on<'a>(&self, update: &'a EventUpdate) -> Option<&'a CommandArgs>

Get the command update args if the update is for this command and scope.

pub fn on_unhandled<'a>( &self, update: &'a EventUpdate, ) -> Option<&'a CommandArgs>

Get the event update args if the update is for this event and propagation is not stopped.

pub fn handle<R>( &self, update: &EventUpdate, handler: impl FnOnce(&CommandArgs) -> R, ) -> Option<R>

Calls handler if the update is for this event and propagation is not stopped, after the handler is called propagation is stopped.

pub fn has_handlers(&self) -> ReadOnlyArcVar<bool>

Gets a variable that tracks if this command has any handlers.

pub fn is_enabled(&self) -> ReadOnlyArcVar<bool>

Gets a variable that tracks if this command has any enabled handlers.

pub fn has_handlers_value(&self) -> bool

Gets if the command has handlers without creating a tracking variable for the state.

pub fn is_enabled_value(&self) -> bool

Gets if the command is enabled without creating a tracking variable for the state.

pub fn visit_scopes<T>( &self, visitor: impl FnMut(Command) -> ControlFlow<T>, ) -> Option<T>

Calls visitor for each scope of this command.

Note that scoped commands are removed if unused, see with_meta for more details.

pub fn notify(&self)

Schedule a command update without param.

pub fn notify_descendants(&self, parent: &WidgetInfo)

Schedule a command update without param for all scopes inside parent.

pub fn notify_param(&self, param: impl Any + Send + Sync)

Schedule a command update with custom param.

pub fn notify_linked( &self, propagation: EventPropagationHandle, param: Option<CommandParam>, )

Schedule a command update linked with an external event propagation.

pub fn on_pre_event<H>(&self, enabled: bool, handler: H) -> EventHandle
where H: AppHandler<AppCommandArgs>,

Creates a preview event handler for the command.

This is similar to Event::on_pre_event, but handler is only called if the command scope matches.

The enabled parameter defines the initial state of the command subscription, the subscription handle is available in the handler args.

pub fn on_event<H>(&self, enabled: bool, handler: H) -> EventHandle
where H: AppHandler<AppCommandArgs>,

Creates an event handler for the command.

This is similar to Event::on_event, but handler is only called if the command scope matches.

The enabled parameter defines the initial state of the command subscription, the subscription handle is available in the handler args.