zng_app::event

Trait AnyEventArgs

pub trait AnyEventArgs: AnyVarValue {
    // Required methods
    fn timestamp(&self) -> DInstant;
    fn propagation(&self) -> &EventPropagationHandle;
    fn is_in_target(&self, widget: WidgetId) -> bool;
    fn clone_boxed(&self) -> Box<dyn AnyEventArgs>;
}

Represents any event update.

Required Methods

fn timestamp(&self) -> DInstant

Instant this event update happened.

fn propagation(&self) -> &EventPropagationHandle

Propagation handle associated with this event instance.

Cloned arguments share the same handle, some arguments may also share the handle of another event if they share the same cause.

fn is_in_target(&self, widget: WidgetId) -> bool

Gets if the widget is in any of the target paths in this update.

fn clone_boxed(&self) -> Box<dyn AnyEventArgs>

Clone the args to a new box.