zng_app::event

Trait CommandNameExt

pub trait CommandNameExt {
    // Required methods
    fn name(self) -> CommandMetaVar<Txt>;
    fn init_name(self, name: impl Into<Txt>) -> Self;
    fn name_with_shortcut(self) -> Var<Txt>
       where Self: CommandShortcutExt;
}

Adds the name command metadata.

Required Methods

fn name(self) -> CommandMetaVar<Txt>

Gets a read-write variable that is the display name for the command.

fn init_name(self, name: impl Into<Txt>) -> Self

Sets the initial name if it is not set.

fn name_with_shortcut(self) -> Var<Txt>
where Self: CommandShortcutExt,

Gets a read-only variable that formats the name and first shortcut in the following format: name (first_shortcut) Note: If no shortcuts are available this method returns the same as name

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.