zng_app::event
Type Alias CommandMetaVar

pub type CommandMetaVar<T> = BoxedVar<T>;

Read-write command metadata variable.

The boxed var is an ArcVar<T> for app scope, or ArcCowVar<T, ArcVar<T>> for scoped commands. If you get this variable from an app scoped command it sets the value for all scopes. If you get this variable using a scoped command, it is a clone-on-write variable that overrides only the value for the scope.

Aliased Type

struct CommandMetaVar<T>(/* private fields */);