zng_app
Module handler

Handler types and macros.

Macros

app_hn

Declare a mutable clone-move app event handler.

app_hn_once

Declare a clone-move app event handler that is only called once.

async_app_hn

Declare an async clone-move app event handler.

async_app_hn_once

Declare an async clone-move app event handler that is only called once.

async_hn

Declare an async clone-move event handler.

async_hn_once

Declare an async clone-move event handler that is only called once.

hn

Declare a mutable clone-move event handler.

hn_once

Declare a clone-move event handler that is only called once.

Structs

AppHandlerArgs

Arguments for a call of AppHandler::event.

FilterAppHandler

App handler wrapper that filters the events, only delegating to self when filter returns true.

FilterWidgetHandler

Widget handler wrapper that filters the events, only delegating to self when filter returns true.

Traits

AppHandler

Represents an event handler in the app context.

AppWeakHandle

Represents a weak handle to an AppHandler subscription.

WidgetHandler

Represents a handler in a widget context.