zng_app

Module render

Frame render and metadata API.

Structs

BuiltFrame

Output of a FrameBuilder.

BuiltFrameUpdate

Output of a FrameBuilder.

ClipBuilder

Builder for a chain of render and hit-test clips.

FontSynthesis

Configure if a synthetic font is generated for fonts that do not implement bold or oblique variants.

FrameBuilder

A full frame builder.

FrameUpdate

A frame quick update.

FrameValueKey

Unique key of an updatable value in the view-process frame.

FrameValueUpdate

Represents an update targeting a previously setup FrameValue.

HitTestBuilder

Builder for the hit-testable shape of the inner-bounds of a widget.

HitTestClipBuilder

Builder for a chain of hit-test clips.

ReferenceFrameId

Represents an unique key for a spatial reference frame that is recreated in multiple frames.

ReuseRange

Represents a display list reuse range.

SpatialFrameId

Unique ID of a reference frame.

Enums

FrameValue

Represents a frame value that may be updated.

ImageRendering

Image scaling algorithm in the renderer.

RepeatMode

Nine-patch border repeat mode.

TransformStyle

Defines if a widget is part of the same 3D space as the parent.

Traits

Font

A text font.

Img

A loaded or loading image.