zng_app

Module shortcut

Key combination types.

This is declared on this crate mostly to support shortcuts in commands, shortcut events are implemented in the input crate.

Macros

shortcut

Creates a Shortcut.

Structs

KeyChord

A sequence of two keyboard combinations.

KeyGesture

A keyboard combination.

ModifiersState

Represents the current state of the keyboard modifiers.

ParseError

Shortcut, gesture parsing error.

ShortcutFilter

Conditions that must be met for the shortcut to apply.

Shortcuts

Multiple shortcuts.

Enums

GestureKey

A keyboard key used in a gesture.

ModifierGesture

A modifier key press and release without any other key press in between.

Shortcut

Keyboard gesture or chord associated with a command.

Traits

CommandShortcutExt

Adds the shortcut metadata.