zng_app
Struct APP

pub struct APP;

Start and manage an app process.

Implementations

impl APP

pub fn exit(&self) -> ResponseVar<ExitCancelled>

Register a request for process exit with code 0 in the next update.

The EXIT_REQUESTED_EVENT will notify, and if propagation is not cancelled the app process will exit.

Returns a response variable that is updated once with the unit value ExitCancelled if the exit operation is cancelled.

See also the EXIT_CMD.

pub fn is_suspended(&self) -> ReadOnlyArcVar<bool>

Gets a variable that tracks if the app is suspended by the operating system.

Suspended apps cannot create graphics contexts and are likely to be killed if the user does not return. Operations that persist data should flush on suspension.

App suspension is controlled by the view-process, the VIEW_PROCESS_SUSPENDED_EVENT notifies on suspension and the VIEW_PROCESS_INITED_EVENT notifies a “respawn” on resume.

impl APP

App time control.

The manual time methods are only recommended for headless apps.

pub fn pause_time_for_update(&self) -> ArcVar<bool>

Gets a variable that configures if INSTANT.now is the same exact value during each update, info, layout or render pass.

Time is paused by default, setting this to false will cause INSTANT.now to read the system time for every call.

pub fn start_manual_time(&self)

Pause the INSTANT.now value, after this call it must be updated manually using advance_manual_time or set_manual_time. To resume normal time use end_manual_time.

pub fn advance_manual_time(&self, advance: Duration)

Adds the advance to the current manual time.

Note that you must ensure an update reaches the code that controls manual time, otherwise the app loop may end-up stuck on idle or awaiting a timer that never elapses.

Panics

Panics if called before start_manual_time.

pub fn set_manual_time(&self, now: DInstant)

Set the current INSTANT.now.

Panics

Panics if called before start_manual_time.

pub fn end_manual_time(&self)

Resumes normal time.

impl APP

pub fn multi_app_enabled(&self) -> bool

If the crate was built with feature="multi_app".

If true multiple apps can run in the same process, but only one app per thread at a time.

pub fn is_running(&self) -> bool

If an app is already running in the current thread.

Apps are running as soon as they start building, and stop running after AppExtended::run returns or the HeadlessApp is dropped.

You can use app_local! to create static resources that live for the app lifetime.

pub fn id(&self) -> Option<AppId>

Gets the unique ID of the current app.

This ID usually does not change as most apps only run once per process, but it can change often during tests. Resources that interact with app_local! values can use this ID to ensure that they are still operating in the same app.

pub fn window_mode(&self) -> WindowMode

Returns a WindowMode value that indicates if the app is headless, headless with renderer or headed.

Note that specific windows can be in headless mode even if the app is headed.

pub fn extensions(&self) -> Arc<AppExtensionsInfo>

List of app extensions that are part of the current app.

pub fn device_events(&self) -> bool

If device events are enabled for the current app.

See AppExtension::enable_device_events for more details.

impl APP

pub fn minimal(&self) -> AppExtended<Vec<Box<dyn AppExtensionBoxed>>>

Starts building an application with no extensions.