zng_app::view_process

Module raw_device_events

Events directly from view-process not targeting any windows.

These events get emitted only if the app APP.device_events_filter allows. When enabled they can be used like raw_events.

Structs

AxisId

Identifier for a specific analog axis on some device.

AxisMotionArgs

Arguments for AXIS_MOTION_EVENT.

ButtonArgs

Arguments for the BUTTON_EVENT.

INPUT_DEVICES

Input devices info service.

InputDeviceCapability

Capabilities of an input device.

InputDeviceId

Unique identifier of an input device event source.

InputDeviceInfo

Info about an human input device.

InputDevicesChangedArgs

Arguments for INPUT_DEVICES_CHANGED_EVENT.

KeyArgs

Arguments for the KEY_EVENT.

PointerMotionArgs

Arguments for POINTER_MOTION_EVENT.

ScrollMotionArgs

Arguments for SCROLL_MOTION_EVENT.

Statics

AXIS_MOTION_EVENT

Motion on some analog axis.

BUTTON_EVENT

Button press/release from a device, probably a mouse.

INPUT_DEVICES_CHANGED_EVENT

Raw input devices have been added or removed from the system.

KEY_EVENT

Keyboard device key press.

POINTER_MOTION_EVENT

Device unfiltered 2D move delta.

SCROLL_MOTION_EVENT

Mouse device unfiltered wheel motion delta.