zng_app::view_process::raw_events
Struct RawKeyInputArgs

pub struct RawKeyInputArgs {
    pub timestamp: DInstant,
    pub window_id: WindowId,
    pub device_id: DeviceId,
    pub key_code: KeyCode,
    pub key_location: KeyLocation,
    pub state: KeyState,
    pub key: Key,
    pub key_modified: Key,
    pub text: Txt,
    /* private fields */
}

Arguments for the RAW_KEY_INPUT_EVENT.

Fields
timestamp: DInstant

Instant the event happened.

window_id: WindowId

Window that received the event.

device_id: DeviceId

Keyboard device that generated the event.

key_code: KeyCode

Physical key.

key_location: KeyLocation

The location of the key on the keyboard.

state: KeyState

If the key was pressed or released.

key: Key

Semantic key.

Pressing Shift+A key will produce Key::Char('a') in QWERTY keyboards, the modifiers are not applied.

key_modified: Key

Semantic key modified by the current active modifiers.

Pressing Shift+A key will produce Key::Char('A') in QWERTY keyboards, the modifiers are applied.

text: Txt

Text typed.

This is only set for KeyState::Pressed of a key that generates text.

This is usually the key_modified char, but is also '\r' for Key::Enter. On Windows when a dead key was pressed earlier but cannot be combined with the character from this key press, the produced text will consist of two characters: the dead-key-character followed by the character resulting from this key press.

Implementations

impl RawKeyInputArgs

pub fn new( timestamp: impl Into<DInstant>, propagation_handle: EventPropagationHandle, window_id: impl Into<WindowId>, device_id: impl Into<DeviceId>, key_code: impl Into<KeyCode>, key_location: impl Into<KeyLocation>, state: impl Into<KeyState>, key: impl Into<Key>, key_modified: impl Into<Key>, text: impl Into<Txt>, ) -> Self

New args from values that convert into the argument types.

pub fn now( window_id: impl Into<WindowId>, device_id: impl Into<DeviceId>, key_code: impl Into<KeyCode>, key_location: impl Into<KeyLocation>, state: impl Into<KeyState>, key: impl Into<Key>, key_modified: impl Into<Key>, text: impl Into<Txt>, ) -> Self

Arguments for event that happened now (INSTANT.now).