zng_app::view_process

Struct VIEW_PROCESS

pub struct VIEW_PROCESS;

Connection to the running view-process for the context app.

Implementations

impl VIEW_PROCESS

pub fn is_available(&self) -> bool

If the VIEW_PROCESS can be used, this is only true in app threads for apps with render, all other methods will panic if called when this is not true.

pub fn is_connected(&self) -> bool

View-process running, connected and ready.

pub fn is_headless_with_render(&self) -> bool

If is running in headless renderer mode.

pub fn is_same_process(&self) -> bool

If is running both view and app in the same process.

pub fn generation(&self) -> ViewProcessGen

Gets the current view-process generation.

pub fn open_window(&self, config: WindowRequest) -> Result<(), ViewChannelError>

Sends a request to open a window and associate it with the window_id.

A RAW_WINDOW_OPEN_EVENT or RAW_WINDOW_OR_HEADLESS_OPEN_ERROR_EVENT will be received in response to this request.

pub fn open_headless( &self, config: HeadlessRequest, ) -> Result<(), ViewChannelError>

Sends a request to open a headless renderer and associate it with the window_id.

Note that no actual window is created, only the renderer, the use of window-ids to identify this renderer is only for convenience.

A RAW_HEADLESS_OPEN_EVENT or RAW_WINDOW_OR_HEADLESS_OPEN_ERROR_EVENT will be received in response to this request.

pub fn add_image( &self, request: ImageRequest<IpcBytes>, ) -> Result<ViewImage, ViewChannelError>

Send an image for decoding.

This function returns immediately, the ViewImage will update when Event::ImageMetadataLoaded, Event::ImageLoaded and Event::ImageLoadError events are received.

pub fn add_image_pro( &self, request: ImageRequest<IpcBytesReceiver>, ) -> Result<ViewImage, ViewChannelError>

Starts sending an image for progressive decoding.

This function returns immediately, the ViewImage will update when Event::ImageMetadataLoaded, Event::ImagePartiallyLoaded, Event::ImageLoaded and Event::ImageLoadError events are received.

pub fn clipboard(&self) -> Result<&ViewClipboard, ViewChannelError>

View-process clipboard methods.

pub fn image_decoders(&self) -> Result<Vec<Txt>, ViewChannelError>

Returns a list of image decoders supported by the view-process backend.

Each string is the lower-case file extension.

pub fn image_encoders(&self) -> Result<Vec<Txt>, ViewChannelError>

Returns a list of image encoders supported by the view-process backend.

Each string is the lower-case file extension.

pub fn pending_frames(&self) -> usize

Number of frame send that have not finished rendering.

This is the sum of pending frames for all renderers.

pub fn respawn(&self)

Reopen the view-process, causing another Event::Inited.

pub fn extension_id( &self, extension_name: impl Into<ApiExtensionName>, ) -> Result<Option<ApiExtensionId>, ViewChannelError>

Gets the ID for the extension_name in the current view-process.

The ID can change for every view-process instance, you must subscribe to the VIEW_PROCESS_INITED_EVENT to refresh the ID. The view-process can respawn at any time in case of error.

pub fn third_party_licenses(&self) -> Result<Vec<LicenseUsed>, ViewChannelError>

Licenses that may be required to be displayed in the app about screen.

This is specially important for prebuilt view users, as the tools that scrap licenses may not find the prebuilt dependencies.

pub fn app_extension_raw( &self, extension_id: ApiExtensionId, extension_request: ApiExtensionPayload, ) -> Result<ApiExtensionPayload, ViewChannelError>

Call an extension with custom encoded payload.

pub fn app_extension<I, O>( &self, extension_id: ApiExtensionId, request: &I, ) -> Result<Result<O, ApiExtensionRecvError>, ViewChannelError>
where I: Serialize, O: DeserializeOwned,

Call an extension with payload request.

pub fn handle_disconnect(&self, vp_gen: ViewProcessGen)

Handle an Event::Disconnected.

The process will exit if the view-process was killed by the user.