zng_app::view_process

Struct ViewImage

pub struct ViewImage(/* private fields */);

Handle to an image loading or loaded in the View Process.

The image is disposed when all clones of the handle are dropped.

Implementations

impl ViewImage

pub fn id(&self) -> Option<ImageId>

Image id.

pub fn is_dummy(&self) -> bool

If the image does not actually exists in the view-process.

pub fn is_loaded(&self) -> bool

Returns true if the image has successfully decoded.

pub fn is_partially_loaded(&self) -> bool

Returns true if the image is progressively decoding and has partially decoded.

pub fn is_error(&self) -> bool

if error is Some.

pub fn error(&self) -> Option<Txt>

Returns the load error if one happened.

pub fn size(&self) -> PxSize

Returns the pixel size, or zero if is not loaded or error.

pub fn partial_size(&self) -> PxSize

Actual size of the current pixels.

Can be different from size if the image is progressively decoding.

pub fn ppi(&self) -> Option<ImagePpi>

Returns the โpixels-per-inchโ metadata associated with the image, or None if not loaded or error or no metadata provided by decoder.

pub fn is_opaque(&self) -> bool

Returns if the image is fully opaque.

pub fn is_mask(&self) -> bool

Returns if the image is a single channel mask (A8).

pub fn partial_pixels(&self) -> Option<Vec<u8>>

Copy the partially decoded pixels if the image is progressively decoding and has not finished decoding.

Format is BGRA8 for normal images or A8 if is_mask.

pub fn pixels(&self) -> Option<IpcBytes>

Reference the decoded pixels of image.

Returns None until the image is fully loaded. Use partial_pixels to copy partially decoded bytes.

Format is pre-multiplied BGRA8 for normal images or A8 if is_mask.

pub fn app_id(&self) -> Option<AppId>

Returns the app that owns the view-process that is handling this image.

pub fn generation(&self) -> ViewProcessGen

Returns the view-process generation on which the image is loaded.

pub fn downgrade(&self) -> WeakViewImage

Creates a WeakViewImage.

pub fn dummy(error: Option<Txt>) -> Self

Create a dummy image in the loaded or error state.

pub fn awaiter(&self) -> SignalOnce

Returns a future that awaits until this image is loaded or encountered an error.

pub async fn encode(&self, format: Txt) -> Result<IpcBytes, EncodeError>

Tries to encode the image to the format.

The format must be one of the image_encoders supported by the view-process backend.