Struct zng_app::view_process::ViewWindow

pub struct ViewWindow(/* private fields */);

Handle to a window open in the view-process.

The window is closed when all clones of the handle are dropped.

Implementations

impl ViewWindow

pub fn generation(&self) -> ViewProcessGen

Returns the view-process generation on which the window was open.

pub fn set_title(&self, title: Txt) -> Result<(), ViewProcessOffline>

Set the window title.

pub fn set_visible(&self, visible: bool) -> Result<(), ViewProcessOffline>

Set the window visibility.

pub fn set_always_on_top( &self, always_on_top: bool, ) -> Result<(), ViewProcessOffline>

Set if the window is “top-most”.

pub fn set_movable(&self, movable: bool) -> Result<(), ViewProcessOffline>

Set if the user can drag-move the window.

pub fn set_resizable(&self, resizable: bool) -> Result<(), ViewProcessOffline>

Set if the user can resize the window.

pub fn set_icon( &self, icon: Option<&ViewImage>, ) -> Result<(), ViewProcessOffline>

Set the window icon.

pub fn set_cursor( &self, cursor: Option<CursorIcon>, ) -> Result<(), ViewProcessOffline>

Set the window cursor icon and visibility.

pub fn set_cursor_image( &self, cursor: Option<&ViewImage>, hotspot: PxPoint, ) -> Result<(), ViewProcessOffline>

Set the window cursor to a custom image.

Falls back to cursor icon if set to None.

The hotspot value is an exact point in the image that is the mouse position. This value is only used if the image format does not contain a hotspot.

pub fn set_taskbar_visible( &self, visible: bool, ) -> Result<(), ViewProcessOffline>

Set the window icon visibility in the taskbar.

pub fn bring_to_top(&self) -> Result<(), ViewProcessOffline>

Bring the window the z top.

pub fn set_state(&self, state: WindowStateAll) -> Result<(), ViewProcessOffline>

Set the window state.

pub fn set_video_mode(&self, mode: VideoMode) -> Result<(), ViewProcessOffline>

Set video mode used in exclusive fullscreen.

pub fn set_enabled_buttons( &self, buttons: WindowButton, ) -> Result<(), ViewProcessOffline>

Set enabled window chrome buttons.

pub fn renderer(&self) -> ViewRenderer

Reference the window renderer.

pub fn set_capture_mode(&self, enabled: bool) -> Result<(), ViewProcessOffline>

Sets if the headed window is in capture-mode. If true the resources used to capture a screenshot may be kept in memory to be reused in the next screenshot capture.

Note that capture must still be requested in each frame request.

pub fn focus(&self) -> Result<FocusResult, ViewProcessOffline>

Brings the window to the front and sets input focus.

This request can steal focus from other apps disrupting the user, be careful with it.

pub fn set_focus_indicator( &self, indicator: Option<FocusIndicator>, ) -> Result<(), ViewProcessOffline>

Sets the user attention request indicator, the indicator is cleared when the window is focused or if canceled by setting to None.

pub fn drag_move(&self) -> Result<(), ViewProcessOffline>

Moves the window with the left mouse button until the button is released.

There’s no guarantee that this will work unless the left mouse button was pressed immediately before this function is called.

pub fn drag_resize( &self, direction: ResizeDirection, ) -> Result<(), ViewProcessOffline>

Resizes the window with the left mouse button until the button is released.

There’s no guarantee that this will work unless the left mouse button was pressed immediately before this function is called.

pub fn open_title_bar_context_menu( &self, position: DipPoint, ) -> Result<(), ViewProcessOffline>

Open system title bar context menu.

pub fn message_dialog( &self, dlg: MsgDialog, responder: ResponderVar<MsgDialogResponse>, ) -> Result<(), ViewProcessOffline>

Shows a native message dialog for the window.

The window is not interactive while the dialog is visible and the dialog may be modal in the view-process. In the app-process this is always async, and the response var will update once when the user responds.

pub fn file_dialog( &self, dlg: FileDialog, responder: ResponderVar<FileDialogResponse>, ) -> Result<(), ViewProcessOffline>

Shows a native file/folder dialog for the window.

The window is not interactive while the dialog is visible and the dialog may be modal in the view-process. In the app-process this is always async, and the response var will update once when the user responds.

pub fn access_update( &self, update: AccessTreeUpdate, ) -> Result<(), ViewProcessOffline>

Update the window’s accessibility info tree.

pub fn set_ime_area( &self, area: Option<DipRect>, ) -> Result<(), ViewProcessOffline>

Enable or disable IME by setting a cursor area.

In mobile platforms also shows the software keyboard for Some(_) and hides it for None.

pub fn set_system_shutdown_warn( &self, reason: Txt, ) -> Result<(), ViewProcessOffline>

Attempt to set a system wide shutdown warning associated with the window.

Operating systems that support this show the reason in a warning for the user, it must be a short text that identifies the critical operation that cannot be cancelled.

Note that there is no guarantee that the view-process or operating system will actually set a block, there is no error result because operating systems can silently ignore block requests at any moment, even after an initial successful block.

Set to an empty text to remove the warning.

pub fn close(self)

Drop self.

pub fn window_extension_raw( &self, extension_id: ApiExtensionId, request: ApiExtensionPayload, ) -> Result<ApiExtensionPayload, ViewProcessOffline>

Call a window extension with custom encoded payload.

pub fn window_extension<I, O>( &self, extension_id: ApiExtensionId, request: &I, ) -> Result<Result<O, ApiExtensionRecvError>, ViewProcessOffline>
where I: Serialize, O: DeserializeOwned,

Call a window extension with serialized payload.