zng_app::widget

Attribute Macro easing

#[easing]

Expands a property assign to include a build action that applies an easing transition to the variable inputs.

Full Documentation

Read the documentation in the zng::widget::easing.