zng_app::widget

Module base

The widget base, nodes and properties used in most widgets.

Modules

node

Basic nodes for widgets, some used in WidgetBase.

Structs

NonWidgetBase

Base widget for types that build to a custom type that is not used as a part of the UI tree.

Parallel

Node list methods that are made parallel.

WidgetBase

Base widget.

Enums

HitTestMode

Defines if and how a widget is hit-tested.

Statics

HIT_TEST_MODE_VAR

Defines the hit-test mode for a widget and descendants.

PARALLEL_VAR

Defines what node list methods can run in parallel in a widget and descendants.

Traits

WidgetImpl

Trait implemented by all #[widget].

Functions

id

c Unique ID of the widget instance.