zng_app::widget::base

Enum HitTestMode

pub enum HitTestMode {
    Disabled,
    Bounds,
    RoundedBounds,
    Detailed,
}

Defines if and how a widget is hit-tested.

Variants

Disabled

Widget is never hit.

This mode affects the entire UI branch, if set it disables hit-testing for the widget and all its descendants, even if they set explicitly set their hit-test mode to something else.

Bounds

Widget is hit by any point that intersects the transformed inner bounds rectangle. If the widget is inlined excludes the first row advance and the last row trailing space.

RoundedBounds

Same as Bounds, but also excludes the outside of rounded corners.

This is the default mode.

Detailed

Widget is hit by any point that intersects the hit-test shape defined on render by FrameBuilder::hit_test and auto hit-test.

Implementations

impl HitTestMode

pub fn is_hit_testable(&self) -> bool

Returns true if is any mode other then Disabled.