zng_app::widget::base

Function id

pub fn id(wgt: &mut WidgetBuilding, id: impl IntoValue<WidgetId>)

B Unique ID of the widget instance.

Note that the id can convert from a &'static str unique name.

Mixin Property

This property is a mixin, it modifies the widget builder during build, it does not produce a standalone UI node.