zng_app::widget::base
Module node

Basic nodes for widgets, some used in WidgetBase.

Functions

build

Capture the id property and builds the base widget.

include_intrinsics

Insert widget_child and widget_inner in the widget.

widget

Create a widget node that wraps child and introduces a new widget context. The node defines an WIDGET context and implements the widget in each specific node method.

widget_child

Returns a node that wraps child and potentially applies child transforms if the child turns out to not be a full widget or to be multiple children. This is important for making properties like padding or content_align work for any UiNode as content.

widget_inner

Returns a node that wraps child and marks the WidgetLayout::with_inner and FrameBuilder::push_inner.