zng_app::widget::base::node

Function build

pub fn build(wgt: WidgetBuilder) -> impl UiNode

Capture the id property and builds the base widget.

Note that this function does not handle missing child node, it falls back to FillUiNode. The WidgetBase widget uses the FillUiNode if none was set.