zng_app::widget::base::node

Function widget_child

pub fn widget_child(child: impl IntoUiNode) -> UiNode

Returns a node that wraps child and potentially applies child transforms if the child turns out to not be a full widget or to be multiple children. This is important for making properties like padding or content_align work for any UiNode as content.

This node also pass through the child inline layout return info if the widget and child are inlining and the widget has not set inline info before delegating measure.

This node must be intrinsic at NestGroup::CHILD, the WidgetBase default intrinsic inserts it.