zng_app::widget::base::node

Function widget_inner

pub fn widget_inner(child: impl UiNode) -> impl UiNode

Returns a node that wraps child and marks the WidgetLayout::with_inner and FrameBuilder::push_inner.

This node renders the inner transform and implements the HitTestMode for the widget.

This node must be intrinsic at NestGroup::BORDER, the WidgetBase default intrinsic inserts it.