Struct zng_app::widget::base::NonWidgetBase

pub struct NonWidgetBase { /* private fields */ }

Base widget for types that build to a custom type that is not used as a part of the UI tree.

This type can be used as base instead of WidgetBase for types that provide a custom build that outputs an instance that is not used as a widget in the UI tree.

Implementations

impl NonWidgetBase

pub fn widget_type() -> WidgetType

Gets the type of NonWidgetBase.

pub fn widget_new() -> Self

Starts building a new NonWidgetBase instance.

pub fn widget_builder(&mut self) -> &mut WidgetBuilder

Returns a mutable reference to the widget builder.

pub fn widget_when(&mut self) -> Option<&mut WhenInfo>

Returns a mutable reference to the when block if called inside a when block.

pub fn widget_take(&mut self) -> WidgetBuilder

Takes the widget builder, finishing the widget macro build.

After this call trying to set a property using self will panic, the returned builder can still be manipulated directly.

pub fn widget_build(&mut self) -> WidgetBuilder

Finishes the build.

This is the fallback build that simply returns the builder, inheritors should override this method.

pub fn widget_importance(&mut self) -> &mut Importance

Returns a mutable reference to the importance of the next property assigns, unsets or when blocks.

Note that during the widget_intrinsic call this is Importance::WIDGET and after it is Importance::INSTANCE.

pub fn start_when_block( &mut self, inputs: Box<[WhenInput]>, state: BoxedVar<bool>, expr: &'static str, location: SourceLocation, )

Start building a when block, all properties set after this call are pushed in the when block.

pub fn end_when_block(&mut self)

End the current when block, all properties set after this call are pushed in the widget.