zng_app::widget::border

Enum BorderStyle

#[repr(u8)]
pub enum BorderStyle {
    Solid = 1,
    Double = 2,
    Dotted = 3,
    Dashed = 4,
    Hidden = 5,
    Groove = 6,
    Ridge = 7,
    Inset = 8,
    Outset = 9,
}

The line style for the sides of a widgetโs border.

Variants

Solid = 1

Displays a single, straight, solid line.

Double = 2

Displays two straight lines that add up to the pixel size defined by the side width.

Dotted = 3

Displays a series of rounded dots.

Dashed = 4

Displays a series of short square-ended dashes or line segments.

Hidden = 5

Fully transparent line.

Groove = 6

Displays a border with a carved appearance.

Ridge = 7

Displays a border with an extruded appearance.

Inset = 8

Displays a border that makes the widget appear embedded.

Outset = 9

Displays a border that makes the widget appear embossed.