zng_app::widget

Module builder

Widget and property builder types.

Macros

property_args

New Box<PropertyArgs> box from a property and value.

property_id

New PropertyId that represents the type and name.

property_info

New PropertyInfo from property path.

property_input_types

Gets the strong input storage types from a property path.

source_location

New SourceLocation that represents the location you call this macro.

widget_type

Gets the WidgetType info of a widget.

Structs

AnyWhenArcWidgetHandlerBuilder

A when builder for AnyArcWidgetHandler values.

ArcWidgetHandler

Represents a WidgetHandler<A> that can be reused.

BuilderProperty

Represents a property removed from WidgetBuilding.

BuilderPropertyMut

Represents a mutable reference to property in WidgetBuilder or WidgetBuilding.

BuilderPropertyRef

Represents a property in WidgetBuilder or WidgetBuilding.

Importance

Value that indicates the override importance of a property instance, higher overrides lower.

NestGroup

Property nest position group.

NestPosition

Represents the sort index of a property or intrinsic node in a widget instance.

PropertyBuildAction

Represents a custom build action targeting a property input that is applied after when is build.

PropertyBuildActionArgs

Arguments for PropertyBuildAction<I>.

PropertyId

Unique ID of a property implementation.

PropertyInfo

Property info.

PropertyInput

Property input info.

PropertyInputTypes

Represents the strong types of each input of a property.

PropertyNewArgs

Args for PropertyInfo::new closure.

SourceLocation

A location in source-code.

UiNodeInWhenExprError

Error value used in a reference to an UiNode property input is made in when expression.

UiNodeListInWhenExprError

Error value used in a reference to an UiNodeList property input is made in when expression.

WhenBuildAction

Data for a custom when build action associated with an WhenInfo.

WhenInfo

Represents a when block in a widget.

WhenInput

Input var read in a when condition expression.

WhenInputVar

Represents a WhenInput variable that can be rebound.

WidgetBuilder

Widget instance builder.

WidgetBuilderProperties

Direct property access in WidgetBuilder and WidgetBuilding.

WidgetBuilding

Represents a finalizing WidgetBuilder.

WidgetHandlerInWhenExprError

Error value used in a reference to an UiNodeList property input is made in when expression.

WidgetType

Unique identifier of a widget type.

Enums

InputKind

Kind of property input.

WhenInputMember

Represents what member and how it was accessed in a WhenInput.

Traits

AnyArcWidgetHandler

Represents a type erased ArcWidgetHandler<A>.

AnyPropertyBuildAction

Represents any PropertyBuildAction<I>.

PropertyArgs

Represents a property instantiation request.

Type Aliases

PropertyBuildActions

Property build actions that must be applied to property args.

PropertyBuildActionsWhenData

Data for property build actions associated with when conditions.