zng_app::widget::builder

Macro property_input_types

macro_rules! property_input_types {
    ($($tt:tt)*) => { ... };
}

Gets the strong input storage types from a property path.

See PropertyInputTypes<Tuple> for more details.

Syntax

property::path: Gets the input types for the property function.

Self::property: Gets the input types for the property method on the widget.