zng_app::widget::builder
Struct ArcWidgetHandler

pub struct ArcWidgetHandler<A: Clone + 'static>(/* private fields */);

Represents a WidgetHandler<A> that can be reused.

Note that hn_once! will still only be used once, and async_hn! tasks are bound to the specific widget context that spawned them. This struct is cloneable to support handler properties in styleable widgets, but the general expectation is that the handler will be used on one property instance at a time.

Implementations

impl<A: Clone + 'static> ArcWidgetHandler<A>

pub fn new(handler: impl WidgetHandler<A>) -> Self

New from handler.