Struct zng_app::widget::builder::WidgetBuilding

pub struct WidgetBuilding { /* private fields */ }

Represents a finalizing WidgetBuilder.

Widgets can register a build action to get access to this, it provides an opportunity to remove or capture the final properties of a widget, after they have all been resolved and when assigns generated. Build actions can also define the child node, intrinsic nodes and a custom builder.

Implementations

impl WidgetBuilding

pub fn widget_type(&self) -> WidgetType

The widget that started this builder.

pub fn has_child(&self) -> bool

If an innermost node is defined.

If false by the end of build the FillUiNode is used as the innermost node.

pub fn set_child(&mut self, node: impl UiNode)

Set/replace the innermost node of the widget.

pub fn disable_inspector(&mut self)

Don’t insert the inspector node and inspector metadata on build.

The inspector metadata is inserted by default when feature="inspector" is active.

pub fn disable_trace_widget(&mut self)

Don’t insert the widget trace node on build.

The trace node is inserted by default when feature="trace_widget" is active.

pub fn disable_trace_wgt_item(&mut self)

Don’t insert property/intrinsic trace nodes on build.

The trace nodes is inserted by default when feature="trace_wgt_item" is active.

pub fn push_intrinsic<I: UiNode>( &mut self, group: NestGroup, name: &'static str, intrinsic: impl FnOnce(BoxedUiNode) -> I + Send + Sync + 'static, )

Insert intrinsic node, that is a core functionality node of the widget that cannot be overridden.

The name is used for inspector/trace only, intrinsic nodes are not deduplicated.

pub fn push_intrinsic_positioned<I: UiNode>( &mut self, position: NestPosition, name: &'static str, intrinsic: impl FnOnce(BoxedUiNode) -> I + Send + Sync + 'static, )

Insert intrinsic node with custom nest position.

The name is used for inspector/trace only, intrinsic nodes are not deduplicated.

pub fn remove_property( &mut self, property_id: PropertyId, ) -> Option<BuilderProperty>

Removes the property.

Note that if the property can already be captured by another widget component.

pub fn capture_property( &mut self, property_id: PropertyId, ) -> Option<BuilderPropertyRef<'_>>

Flags the property as captured and returns a reference to it.

Note that captured properties are not instantiated in the final build, but they also are not removed like unset. A property can be “captured” more then once, and if the "inspector" feature is enabled they can be inspected.

pub fn capture_var<T>(&mut self, property_id: PropertyId) -> Option<BoxedVar<T>>
where T: VarValue,

Flags the property as captured and downcast the input var.

pub fn capture_var_or_else<T, F>( &mut self, property_id: PropertyId, or_else: impl FnOnce() -> F, ) -> BoxedVar<T>
where T: VarValue, F: IntoVar<T>,

Flags the property as captured and downcast the input var, or calls or_else to generate a fallback.

pub fn capture_var_or_default<T>( &mut self, property_id: PropertyId, ) -> BoxedVar<T>
where T: VarValue + Default,

Flags the property as captured and downcast the input var, returns a new one with the default value.

pub fn capture_ui_node( &mut self, property_id: PropertyId, ) -> Option<BoxedUiNode>

Flags the property as captured and get the input node.

pub fn capture_ui_node_or_else<F>( &mut self, property_id: PropertyId, or_else: impl FnOnce() -> F, ) -> BoxedUiNode
where F: UiNode,

Flags the property as captured and get the input node, or calls or_else to generate a fallback node.

pub fn capture_ui_node_list( &mut self, property_id: PropertyId, ) -> Option<BoxedUiNodeList>

Flags the property as captured and get the input list.

pub fn capture_ui_node_list_or_else<F>( &mut self, property_id: PropertyId, or_else: impl FnOnce() -> F, ) -> BoxedUiNodeList
where F: UiNodeList,

Flags the property as captured and get the input list, or calls or_else to generate a fallback list.

pub fn capture_ui_node_list_or_empty( &mut self, property_id: PropertyId, ) -> BoxedUiNodeList

Flags the property as captured and get the input list, or returns an empty list.

pub fn capture_widget_handler<A: Clone + 'static>( &mut self, property_id: PropertyId, ) -> Option<ArcWidgetHandler<A>>

Flags the property as captured and downcast the input handler.