zng_app::widget::builder

Trait PropertyArgs

pub trait PropertyArgs: Send + Sync {
    // Required methods
    fn clone_boxed(&self) -> Box<dyn PropertyArgs>;
    fn property(&self) -> PropertyInfo;
    fn instantiate(&self, child: UiNode) -> UiNode;

    // Provided methods
    fn var(&self, i: usize) -> &AnyVar { ... }
    fn value(&self, i: usize) -> &dyn AnyVarValue { ... }
    fn ui_node(&self, i: usize) -> &ArcNode { ... }
    fn widget_handler(&self, i: usize) -> &dyn AnyArcWidgetHandler { ... }
}

Represents a property instantiation request.

Required Methods

fn clone_boxed(&self) -> Box<dyn PropertyArgs>

Clones the arguments.

fn property(&self) -> PropertyInfo

Property info.

fn instantiate(&self, child: UiNode) -> UiNode

Create a property instance with args clone or taken.

If the property is PropertyInfo::capture the child is returned.

Provided Methods

fn var(&self, i: usize) -> &AnyVar

Gets a InputKind::Var.

fn value(&self, i: usize) -> &dyn AnyVarValue

Gets a InputKind::Value.

fn ui_node(&self, i: usize) -> &ArcNode

Gets a InputKind::UiNode.

fn widget_handler(&self, i: usize) -> &dyn AnyArcWidgetHandler

Gets a InputKind::WidgetHandler.

Is a ArcWidgetHandler<A>.

Implementations

impl dyn PropertyArgs + '_

pub fn id(&self) -> PropertyId

Unique ID.

pub fn downcast_value<T>(&self, i: usize) -> &T
where T: VarValue,

Gets a strongly typed value.

Panics if the type does not match.

pub fn downcast_var<T>(&self, i: usize) -> Var<T>
where T: VarValue,

Gets a strongly typed var.

Panics if the variable value type does not match.

pub fn downcast_handler<A>(&self, i: usize) -> &ArcWidgetHandler<A>
where A: 'static + Clone,

Gets a strongly typed widget_handler.

Panics if the args type does not match.

pub fn live_debug(&self, i: usize) -> Var<Txt>

Gets the property input as a debug variable.

If the input is a variable the returned variable will update with it, if not it is a static print.

Note that you must call this in the widget context to get the correct value.

pub fn debug(&self, i: usize) -> Txt

Gets the property input current value as a debug text.

Note that you must call this in the widget context to get the correct value.

pub fn new_build( &self, build_actions: PropertyBuildActions, build_actions_when_data: PropertyBuildActionsWhenData, ) -> Box<dyn PropertyArgs>

Call new with the same instance info and args, but with the build_actions and build_actions_when_data.