Module zng_app::widget::info

Widget info tree.

Re-exports

pub use iter::TreeFilter;

Modules

access

Accessibility metadata types.

iter

Widget info tree iterators.

Structs

HitInfo

A hit-test hit.

HitTestInfo

A hit-test result.

InlineSegmentInfo

Info about a segment in the first or last row of an inlined widget.

InteractionPath

Represents a WidgetPath annotated with each widget’s Interactivity.

Interactivity

Represents the level of interaction allowed for a widget.

InteractivityChangedArgs

INTERACTIVITY_CHANGED_EVENT args.

InteractivityFilterArgs

Argument for a interactivity filter function.

ParallelBuilder

Represents a builder split from the main builder that can be used in parallel and then folded back onto the main builder.

TransformChangedArgs

TRANSFORM_CHANGED_EVENT args.

VisibilityChangedArgs

VISIBILITY_CHANGED_EVENT args.

WidgetBorderInfo

Shared reference to the combined border and corner radius of a WidgetInfo.

WidgetBoundsInfo

Shared reference to layout size, offsets, rendered transforms and bounds of a widget.

WidgetDescendantsRange

Represents the descendants of a widget, allows checking if widgets are descendant with O(1) time.

WidgetInfo

Reference to a widget info in a WidgetInfoTree.

WidgetInfoBuilder

Widget info tree builder.

WidgetInfoChangedArgs

WIDGET_INFO_CHANGED_EVENT args.

WidgetInfoTree

A tree of WidgetInfo.

WidgetInfoTreeStats

Stats over the lifetime of a widget info tree.

WidgetInlineInfo

Info about the inlined rows of the widget.

WidgetInlineMeasure

Info about the input inline connecting rows of the widget.

WidgetLayout

Represents the in-progress layout pass for a widget tree.

WidgetMeasure

Represents the in-progress measure pass for a widget tree.

WidgetPath

Full address of a widget.

Enums

RelativeHitZ

Hit-test result on a widget relative to it’s descendants.

Visibility

Widget visibility.

WidgetInfoMeta

Tag for the WidgetInfo::meta state-map.

Statics

INTERACTIVITY_CHANGED_EVENT

Widget interactivity has changed after an info update.

TRANSFORM_CHANGED_EVENT

A widget global inner transform has changed after render.

VISIBILITY_CHANGED_EVENT

Widget visibility has changed after render.

WIDGET_INFO_CHANGED_EVENT

A window widget tree was rebuild.