Struct zng_app::widget::info::access::WidgetAccessInfoBuilder

pub struct WidgetAccessInfoBuilder<'a> { /* private fields */ }

Accessibility metadata.

Implementations

impl<'a> WidgetAccessInfoBuilder<'a>

pub fn set_role(&mut self, role: AccessRole)

Set the accessibility role of the widget.

pub fn push_command(&mut self, cmd: AccessCmdName)

Add a supported access command.

pub fn set_auto_complete(&mut self, mode: AutoComplete)

Set how input text triggers display of one or more predictions of the user’s intended value for a ComboBox, SearchBox, or TextInput.

pub fn set_checked(&mut self, checked: Option<bool>)

If the widget is checked (Some(true)), unchecked (Some(false)), or if the checked status is indeterminate (None).

pub fn set_current(&mut self, kind: CurrentKind)

Indicates that the widget represents the current item of a kind.

pub fn set_error_message(&mut self, invalid_wgt: impl Into<WidgetId>)

Indicates that the widget is an error message for the invalid_wgt.

The other widget must set_invalid.

pub fn set_active_descendant(&mut self, descendant: impl Into<WidgetId>)

Identifies the currently active widget when focus is on a composite widget.

pub fn set_expanded(&mut self, expanded: bool)

Indicate that the widget toggles the visibility of related widgets.

Use push_controls, or push_owns to indicate the widgets that change visibility based on this value.

pub fn set_popup(&mut self, popup: Popup)

Indicates the availability and type of interactive popup widget.

pub fn set_invalid(&mut self, error: Invalid)

Indicates that the widget’s data is invalid with optional kinds of errors.

pub fn set_label(&mut self, name: impl Into<Txt>)

Sets a custom name for the widget in accessibility info.

Note that if this is not set the WidgetId::name of the widget is used.

pub fn set_level(&mut self, hierarchical_level: NonZeroU32)

Sets the hierarchical level of the widget within a parent scope.

pub fn flag_multi_selectable(&mut self)

Indicates that the user may select more than one item from the current selectable descendants.

pub fn set_orientation(&mut self, orientation: Orientation)

Indicates whether the widget’s orientation is horizontal, vertical, or unknown/ambiguous.

pub fn set_placeholder(&mut self, placeholder: impl Into<Txt>)

Short hint (a word or short phrase) intended to help the user with data entry when a form control has no value.

pub fn flag_read_only(&mut self)

Indicates that the widget is not editable, but is otherwise operable.

pub fn flag_required(&mut self)

Indicates that user input is required on the widget before a form may be submitted.

pub fn flag_selected(&mut self)

Indicates that the widget is selected.

pub fn set_sort(&mut self, direction: SortDirection)

Sets the sort direction for the table or grid items.

pub fn set_value_max(&mut self, max: f64)

Set the maximum value (inclusive).

pub fn set_value_min(&mut self, min: f64)

Set the minimum value (inclusive).

pub fn set_value(&mut self, value: f64)

Set the current value.

pub fn set_value_text(&mut self, value: impl Into<Txt>)

Set a text that is a readable version of the current value.

pub fn set_live(&mut self, indicator: LiveIndicator, atomic: bool, busy: bool)

Indicate that the widget can change, how the change can be announced, if atomic the entire widget must be re-read, if busy the screen reader must wait until the change completes.

pub fn set_col_count(&mut self, count: usize)

Sets the total number of columns in a Table, Grid, or TreeGrid when not all columns are present in tree.

The value 0 indicates that not all columns are in the widget and the application cannot determinate the exact number.

pub fn set_col_index(&mut self, index: usize)

Sets the widget’s column index in the parent table or grid.

pub fn set_col_span(&mut self, span: usize)

Sets the number of columns spanned by the widget in the parent table or grid.

pub fn set_row_count(&mut self, count: usize)

Sets the total number of rows in a Table, Grid, or TreeGrid when not all rows are present in tree.

The value 0 indicates that not all rows are in the widget and the application cannot determinate the exact number.

pub fn set_row_index(&mut self, index: usize)

Sets the widget’s row index in the parent table or grid.

pub fn set_row_span(&mut self, span: usize)

Sets the number of rows spanned by the widget in the parent table or grid.

pub fn set_item_count(&mut self, count: usize)

Sets the number of items in the current set of list items or tree items when not all items in the set are present in the tree.

pub fn set_item_index(&mut self, index: usize)

Sets the widget’s number or position in the current set of list items or tree items when not all items are present in the tree.

pub fn flag_modal(&mut self)

Sets if the widget is modal when displayed.

pub fn set_lang(&mut self, lang: LanguageIdentifier)

Defines the language used by screen-readers to read text in this widget and descendants.

pub fn set_scroll_horizontal(&mut self, normal_x: impl IntoVar<Factor>)

Sets the amount scrolled horizontally if allowed.

The normal_x value can be a read-only variable, the variable can be updated without needing to rebuild info for every pixel scrolled, if the view-process requires access info the value is updated every render together with the widget bounds updates.

The value must be normalized in the 0..=1 range, 0 is showing the content leftmost edge, 1 is showing the content the rightmost edge.

pub fn set_scroll_vertical(&mut self, normal_y: impl IntoVar<Factor>)

Sets the amount scrolled vertically if allowed.

The normal_y value can be a read-only variable, the variable can be updated without needing to rebuild info for every pixel scrolled, if the view-process requires access info the value is updated every render together with the widget bounds updates.

The value must be normalized in the 0..=1 range, 0 is showing the content topmost edge, 1 is showing the content the bottommost edge.

pub fn push_controls(&mut self, controlled_id: impl Into<WidgetId>)

Push a widget whose contents or presence are controlled by this widget.

pub fn push_described_by(&mut self, descriptor_id: impl Into<WidgetId>)

Push a widget that describes this widget.

pub fn set_describes(&mut self, target_id: impl Into<WidgetId>)

Set a widget that is described-by this widget.

When access info for the view-process is build this is converted to a described-by entry. Note that only updated widgets are send to the view-process, so if this relation is dynamic you must request info rebuild for the previous and new target_id to ensure they update correctly.

pub fn push_details(&mut self, detail_id: impl Into<WidgetId>)

Push a widget that provide additional information related to this widget.

pub fn push_labelled_by(&mut self, label_id: impl Into<WidgetId>)

Push a widget that provide additional information related to this widget.

pub fn set_labels(&mut self, target_id: impl Into<WidgetId>)

Set a widget that is labelled-by this widget.

When access info for the view-process is build this is converted to a labelled-by entry. Note that only updated widgets are send to the view-process, so if this relation is dynamic you must request info rebuild for the previous and new target_id to ensure they update correctly.

pub fn push_owns(&mut self, owned_id: impl Into<WidgetId>)

Push a widget that is a child of this widget, but is not already a child in the info tree.

pub fn push_flows_to(&mut self, next_id: impl Into<WidgetId>)

Push an option for next widget read that is not the next logical widget already.

pub fn flag_labelled_by_child(&mut self)

Uses the accessible children as labelled_by.

pub fn flag_inaccessible(&mut self)

Exclude the widget and descendants from the view-process and screen readers.

Note that the accessibility info for the widget and descendants is still collected and available in the app-process.

pub fn on_access_build( &mut self, handler: impl Fn(AccessBuildArgs<'_>) + Send + Sync + 'static )

Register a handler that is called every time view-process access info is build from the current widget, the handler can modify the view info.