zng_app::widget::info

Module iter

Widget info tree iterators.

Structs

Ancestors

Iterator over all ancestors of a widget.

Children

Iterator over all children of a widget.

NextSiblings

Iterator over all next siblings of a widget.

PrevSiblings

Iterator over all next siblings of a widget.

RevTreeIter

Reversing tree iterator.

TreeFilterIter

Filtering tree iterator.

TreeIter

Primary implementer of TreeIterator.

Enums

TreeFilter

Widget tree filter selected for a widget in the tree.

Traits

TreeIterator

Iterator that traverses the branches of a widget tree.