Struct zng_app::widget::info::InteractionPath

pub struct InteractionPath { /* private fields */ }

Represents a WidgetPath annotated with each widget’s Interactivity.

Implementations

impl InteractionPath

pub fn new<P: IntoIterator<Item = (WidgetId, Interactivity)>>( window_id: WindowId, path: P ) -> InteractionPath

New custom path.

pub fn new_enabled( window_id: WindowId, path: Arc<Vec<WidgetId>> ) -> InteractionPath

New custom path with all widgets enabled.

pub fn from_enabled(path: WidgetPath) -> InteractionPath

New interactivity path with all widgets enabled.

pub fn as_path(&self) -> &WidgetPath

Dereferences to the path.

pub fn blocked_index(&self) -> Option<usize>

Index of first BLOCKED.

pub fn disabled_index(&self) -> Option<usize>

Index of first DISABLED.

pub fn interaction_path( &self ) -> impl DoubleEndedIterator<Item = Interactivity> + ExactSizeIterator

Interactivity for each widget, root first.

pub fn interactivity_of(&self, widget_id: WidgetId) -> Option<Interactivity>

Search for the interactivity value associated with the widget in the path.

pub fn interactivity(&self) -> Interactivity

Interactivity of the widget.

pub fn zip( &self ) -> impl DoubleEndedIterator<Item = (WidgetId, Interactivity)> + ExactSizeIterator + '_

Zip widgets and interactivity.

pub fn unblocked(self) -> Option<InteractionPath>

Gets the ENABLED or DISABLED part of the path, or none if the widget is blocked at the root.

pub fn enabled(self) -> Option<WidgetPath>

Gets the ENABLED part of the path, or none if the widget is not enabled at the root.

pub fn ancestor_path( &self, ancestor_id: WidgetId ) -> Option<Cow<'_, InteractionPath>>

Make a path to an ancestor id that is contained in the current path.

pub fn shared_ancestor<'a>( &'a self, other: &'a InteractionPath ) -> Option<Cow<'a, InteractionPath>>

Get the inner most widget parent shared by both self and other with the same interactivity.

pub fn root_path(&self) -> Cow<'_, InteractionPath>

Gets a path to the root widget of this path.

pub fn sub_path(&self, widget_id: WidgetId) -> Option<Cow<'_, InteractionPath>>

Gets a sub-path up to widget_id (inclusive), or None if the widget is not in the path.