zng_app::widget::info

Struct InteractivityChangedArgs

pub struct InteractivityChangedArgs {
    pub timestamp: DInstant,
    pub prev_tree: WidgetInfoTree,
    pub tree: WidgetInfoTree,
    pub changed: IdSet<WidgetId>,
    /* private fields */
}

INTERACTIVITY_CHANGED_EVENT args.

Fields
timestamp: DInstant

Instant the event happened.

prev_tree: WidgetInfoTree

Previous tree with old interactivity values.

tree: WidgetInfoTree

New tree with new interactivity values.

changed: IdSet<WidgetId>

All event subscribers that changed interactivity in this info update.

Implementations

impl InteractivityChangedArgs

pub fn new( timestamp: impl Into<DInstant>, propagation_handle: EventPropagationHandle, prev_tree: impl Into<WidgetInfoTree>, tree: impl Into<WidgetInfoTree>, changed: impl Into<IdSet<WidgetId>>, ) -> Self

New args from values that convert into the argument types.

pub fn now( prev_tree: impl Into<WidgetInfoTree>, tree: impl Into<WidgetInfoTree>, changed: impl Into<IdSet<WidgetId>>, ) -> Self

Arguments for event that happened now (INSTANT.now).

impl InteractivityChangedArgs

pub fn prev_interactivity(&self, widget_id: WidgetId) -> Option<Interactivity>

Previous interactivity of this widget.

Returns None if the widget was not in the previous info tree.

pub fn new_interactivity(&self, widget_id: WidgetId) -> Interactivity

New interactivity of the widget.

Panics

Panics if widget_id is not in tree. This method must be called only for changed.

pub fn is_enable(&self, widget_id: WidgetId) -> bool

Widget was disabled or did not exist, now is enabled.

pub fn is_disable(&self, widget_id: WidgetId) -> bool

Widget was enabled or did not exist, now is disabled.

pub fn is_unblock(&self, widget_id: WidgetId) -> bool

Widget was blocked or did not exist, now is unblocked.

pub fn is_block(&self, widget_id: WidgetId) -> bool

Widget was unblocked or did not exist, now is blocked.

pub fn is_vis_enable(&self, widget_id: WidgetId) -> bool

Widget was visually disabled or did not exist, now is visually enabled.

pub fn is_vis_disable(&self, widget_id: WidgetId) -> bool

Widget was visually enabled or did not exist, now is visually disabled.

pub fn enabled_change( &self, widget_id: WidgetId, ) -> Option<(Option<Interactivity>, Interactivity)>

Returns the previous and new interactivity if the widget was enabled, disabled or is new.

pub fn vis_enabled_change( &self, widget_id: WidgetId, ) -> Option<(Option<Interactivity>, Interactivity)>

Returns the previous and new interactivity if the widget was visually enabled, visually disabled or is new.

pub fn blocked_change( &self, widget_id: WidgetId, ) -> Option<(Option<Interactivity>, Interactivity)>

Returns the previous and new interactivity if the widget was blocked, unblocked or is new.

pub fn is_new(&self, widget_id: WidgetId) -> bool

Widget is new, no previous interactivity state is known, events that filter by interactivity change update by default if the widget is new.