zng_app::widget::info

Struct WidgetBorderInfo

pub struct WidgetBorderInfo(/* private fields */);

Shared reference to the combined border and corner radius of a WidgetInfo.

Implementations

impl WidgetBorderInfo

pub fn new() -> Self

New default.

pub fn offsets(&self) -> PxSideOffsets

Sum of the widths of all borders set on the widget.

pub fn corner_radius(&self) -> PxCornerRadius

Corner radius set on the widget, this is the outer curve of border corners.

pub fn inner_corner_radius(&self) -> PxCornerRadius

Computes the corner_radius deflated by offsets, this is the inner curve of border corners.

pub fn inner_offset(&self, bounds: &WidgetBoundsInfo) -> PxVector

Compute the inner offset plus offsets left, top.

pub fn inner_size(&self, bounds: &WidgetBoundsInfo) -> PxSize

Compute the inner size offset by offsets.

pub fn inner_transform(&self, bounds: &WidgetBoundsInfo) -> PxTransform

Compute the inner transform offset by the offsets.