zng_app::widget::info
Struct WidgetBoundsInfo

pub struct WidgetBoundsInfo(/* private fields */);

Shared reference to layout size, offsets, rendered transforms and bounds of a widget.

Can be retrieved in the WIDGET and WidgetInfo.

Implementations

impl WidgetBoundsInfo

pub fn new() -> Self

New default.

pub fn new_size(outer: PxSize, inner: PxSize) -> Self

New info with bound sizes known.

pub fn measure_outer_size(&self) -> PxSize

Gets the widget’s last measured outer bounds size.

This size is expected to be the same if the widget is layout using the same exact parameters it was measured.

pub fn outer_size(&self) -> PxSize

Gets the widget’s last layout outer bounds size.

pub fn inner_offset(&self) -> PxVector

Gets the widget’s inner bounds offset inside the outer bounds.

If inner_offset_baseline is true the baseline is added from this value.

pub fn inner_offset_baseline(&self) -> bool

If the baseline is added from the inner_offset.

pub fn child_offset(&self) -> PxVector

Gets the widget’s child offset inside the inner bounds.

If the widget’s child is another widget this is zero and the offset is set on that child’s parent_child_offset instead.

pub fn inner_size(&self) -> PxSize

Gets the widget’s inner bounds size.

pub fn baseline(&self) -> Px

The baseline offset up from the inner bounds bottom line.

Note that if inner_offset_baseline is true the inner_offset is already added by the baseline. Parent panel widgets implementing baseline offset must use the final_baseline value to avoid offsetting more then once.

pub fn final_baseline(&self) -> Px

Gets the baseline of the widget after inner_offset is applied.

Returns Px(0) if inner_offset_baseline, otherwise returns baseline.

pub fn outer_transform(&self) -> PxTransform

Gets the global transform of the widget’s outer bounds during the last render or render update.

pub fn parent_child_offset(&self) -> PxVector

Offset rendered in the widget inner set by the parent widget.

Note that this offset is applied to the outer_transform already.

pub fn inner_transform(&self) -> PxTransform

Gets the global transform of the widget’s inner bounds during the last render or render update.

pub fn measure_inline(&self) -> Option<WidgetInlineMeasure>

Gets the latest inline measure info.

Note that this info may not be the same that was used to update the inline layout info. This value is only useful for panels implementing inline, just after the widget was measured.

Returns None if the latest widget measure was not in an inlining context.

pub fn inline(&self) -> Option<MappedMutexGuard<'_, WidgetInlineInfo>>

Exclusive read the latest inline layout info.

Returns None if the latest widget layout was not in an inlining context.

pub fn rendered(&self) -> Option<bool>

Gets the widget’s latest render info, if it was rendered visible or hidden. Returns None if the widget was collapsed.

pub fn is_in_bounds(&self) -> bool

Gets if the inner_bounds are fully inside the parent inner bounds.

pub fn can_auto_hide(&self) -> bool

Gets if the widget only renders if outer_bounds intersects with the FrameBuilder::auto_hide_rect.

This is true by default and can be disabled using allow_auto_hide. If set to false the widget is always rendered, but descendant widgets can still auto-hide.

pub fn outer_bounds(&self) -> PxRect

Outer bounding box, updated after every render.

pub fn inner_bounds(&self) -> PxRect

Calculate the bounding box that envelops the actual size and position of the inner bounds last rendered.

pub fn is_collapsed(&self) -> bool

If the widget and descendants was collapsed during layout.

pub fn transform_style(&self) -> TransformStyle

Gets if the widget preserves 3D perspective.

pub fn perspective(&self) -> Option<(f32, PxPoint)>

Gets the widget perspective and perspective origin (in the inner bounds).

pub fn metrics(&self) -> Option<LayoutMetricsSnapshot>

Snapshot of the LayoutMetrics on the last layout.

The metrics_used value indicates what fields where actually used in the last layout.

Is None if the widget is collapsed.

pub fn metrics_used(&self) -> LayoutMask

All metrics fields used by the widget or descendants on the last layout.

pub fn hit_test_z(&self, window_point: PxPoint) -> RelativeHitZ

Gets the relative hit-test Z for window_point against the hit-test shapes rendered for the widget.

pub fn hit_test_clip_child( &self, child: &WidgetInfo, window_point: PxPoint, ) -> bool

Returns true if a hit-test clip that affects the child removes the window_point hit on the child.