zng_app::widget::info

Struct WidgetInfoTreeStats

#[non_exhaustive]
pub struct WidgetInfoTreeStats {
    pub generation: u32,
    pub build_time: Duration,
    pub reused_widgets: u32,
    pub last_frame: FrameId,
    pub bounds_updated_frame: FrameId,
    pub bounds_updated: u32,
    pub vis_updated_frame: FrameId,
}

Stats over the lifetime of a widget info tree.

The stats for a tree are available in WidgetInfoTree::stats.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

generation: u32

Number of times info was rebuild for the window.

build_time: Duration

Duration of the UiNode::info call for the window content.

reused_widgets: u32

Count of widgets that where reused from a previous tree.

last_frame: FrameId

Last window frame that rendered this tree.

Before the first render this is FrameId::INVALID.

bounds_updated_frame: FrameId

Last window frame that moved or resized the inner bounds of at least one widget.

bounds_updated: u32

Count of moved or resized widgets in the last bounds_updated_frame.

vis_updated_frame: FrameId

Last window frame that changed visibility of at least one widget.