Struct zng_app::widget::info::WidgetInlineInfo

pub struct WidgetInlineInfo {
    pub rows: Vec<PxRect>,
    pub first_segs: Vec<InlineSegmentInfo>,
    pub last_segs: Vec<InlineSegmentInfo>,
    pub inner_size: PxSize,
    /* private fields */
}

Info about the inlined rows of the widget.

Fields
rows: Vec<PxRect>

Last layout rows of the widget.

The rectangles are in the widget’s inner space, from top to bottom.

first_segs: Vec<InlineSegmentInfo>

Segments of the first row.

If this is empty the entire row width is a continuous segment, otherwise the row is segmented and the widget can be interleaved with sibling widgets due to Unicode bidirectional text sorting algorithm.

Note that the segment count may be less then WidgetInlineMeasure::first_segs as contiguous segments may be merged.

The segments are from left to right.

last_segs: Vec<InlineSegmentInfo>

Segments of the last row.

inner_size: PxSize

Widget inner size when the rows where last updated.

Implementations

impl WidgetInlineInfo

pub fn set_first_segs(&mut self, segs: impl Iterator<Item = InlineSegmentInfo>)

Replace the first_segs with segs.

The segments are sorted when needed, but prefer inputs that are mostly sorted.

The segments are merged when there is no gap or there is a small one pixel overlap to the previous segment.

pub fn set_last_segs(&mut self, segs: impl Iterator<Item = InlineSegmentInfo>)

Replace the last_segs with segs.

The segments are sorted when needed, but prefer inputs that are mostly sorted.

The segments are merged when there is no gap or there is a small one pixel overlap to the previous segment.

pub fn union(&self) -> PxRect

Gets the union of all row rectangles.

pub fn negative_space(&self) -> Arc<Vec<PxRect>>

Gets or computes the negative space of the rows in the inner_size space, that is, all the areas that are not covered by any row and not covered by the first and last row segments.

This is computed on demand and cached.

pub fn invalidate_negative_space(&mut self)

Invalidates the negative_space cache.

pub fn clear(&mut self)

Return info to default state, but retain memory for reuse.

pub fn is_default(&self) -> bool

If all value are not different from initial.

This indicates the widget has not handled the inline config yet.