zng_app::widget::info

Struct WidgetInlineMeasure

#[non_exhaustive]
pub struct WidgetInlineMeasure {
    pub first: PxSize,
    pub first_wrapped: bool,
    pub first_segs: Arc<Vec<InlineSegment>>,
    pub last: PxSize,
    pub last_wrapped: bool,
    pub last_segs: Arc<Vec<InlineSegment>>,
}

Info about the input inline connecting rows of the widget.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

first: PxSize

Preferred first size.

In left-to-right direction the origin is top_left, in right-to-left direction the origin is top_right - first.width.

first_wrapped: bool

Indicates that first starts in the next row, not in the current row defined by the inline constraints.

first_segs: Arc<Vec<InlineSegment>>

Inline segments in the first row.

The sum of segment widths must be less or equal to the first.width.

last: PxSize

Preferred last size.

In left-to-right direction the origin is bottom_left - last.height, in right-to-left direction the origin is bottom_right - last.

Must be equal to first if did not wrap.

Must not be empty if first is not empty, that is, must not wrap if the last item can fit in the previous row.

last_wrapped: bool

Indicates that last starts in a next row, not in the same row as the first.

last_segs: Arc<Vec<InlineSegment>>

Inline segments in the last row.

The sum of segment widths must be less or equal to the last.width.

Implementations

impl WidgetInlineMeasure

pub fn with_first_segs(&mut self, f: impl FnOnce(&mut Vec<InlineSegment>))

Visit a mutable reference to the new first_segs value, f is called with an empty vec that can be reused or new.

pub fn with_last_segs(&mut self, f: impl FnOnce(&mut Vec<InlineSegment>))

Visit a mutable reference to the new last_segs value, f is called with an empty vec that can be reused or new.

pub fn is_default(&self) -> bool

If all value are not different from initial.

This indicates the widget has not handled the inline config yet.