Struct zng_app::widget::info::WidgetMeasure

pub struct WidgetMeasure { /* private fields */ }

Represents the in-progress measure pass for a widget tree.

Use WidgetLayout::to_measure to instantiate.

Implementations

impl WidgetMeasure

pub fn new_reuse(inline: Option<WidgetInlineMeasure>) -> Self

New with no widget layouts invalidated.

Prefer WidgetLayout::to_measure instead of this.

pub fn is_inline(&self) -> bool

If the parent widget is doing inline flow layout.

pub fn inline(&mut self) -> Option<&mut WidgetInlineMeasure>

Mutable reference to the current widget’s inline info.

The widget must configure this to be inlined in parent layout. This is only Some(_) if inline is enabled.

See WidgetInlineMeasure for more details.

pub fn disable_inline(&mut self)

Sets is_inline to false.

Must be called before child delegation, otherwise children that inline may render expecting to fit in the inline flow.

Note that this disables inline for the calling widget’s next layout too, every property that affects layout and does not support inline layout must propagate measure using this method to correctly configure the widget.

Prefer measure_block as if also clears the layout constraints.

pub fn measure_block(&mut self, child: &mut impl UiNode) -> PxSize

Disable inline and measure child with no inline constraints.

pub fn measure_inline( &mut self, first_max: Px, mid_clear_min: Px, child: &mut impl UiNode ) -> (Option<WidgetInlineMeasure>, PxSize)

Measure the child node with inline enabled for the child node context.

The first_max and mid_clear_min parameters match the InlineConstraintsMeasure members, and will be set in the child context.

Note that this does not enabled inline in the calling widget if inlining was disabled by the parent nodes, it creates a new inlining context.

Returns the inline requirements of the child and its desired bounds size, returns None requirements if the child disables inline or is not a full widget.

pub fn with_widget( &mut self, measure: impl FnOnce(&mut Self) -> PxSize ) -> PxSize

Measure a widget.

pub fn with_inline_visual( &mut self, measure: impl FnOnce(&mut Self) -> PxSize ) -> PxSize

Calls measure with inline force enabled on the widget.

The widget will be inlining even if the parent widget is not inlining, if properties request disable_inline these requests are ignored.

pub fn parallel_split(&self) -> ParallelBuilder<WidgetMeasure>

Start a parallel measure.

Returns an instance that can be used to acquire multiple mutable WidgetMeasure during measure. The parallel_fold method must be called after the parallel processing is done.

pub fn parallel_fold(&mut self, split: ParallelBuilder<WidgetMeasure>)

Collect the parallel changes back.