zng_app::widget

Module node

Widget nodes types, UiNode, UiVec and others.

Macros

ui_vec

Creates an UiVec containing the arguments.

Structs

AdoptiveChildNode

Placeholder for the dynamic child of an adoptive node.

AdoptiveNode

Represents a node setup to dynamically swap child.

ArcNode

A reference counted UiNode.

ChainList

UI node list implementation that flattens child lists.

DefaultPanelListData

Default PanelList associated data.

EditableUiVec

Represents an UiVec that can be modified using a connected sender.

EditableUiVecRef

Represents a sender to an EditableUiVec.

FillUiNode

A UI node that fills the available layout space.

MatchNodeChild

Child node of match_node.

MatchWidgetChild

Child node of match_widget.

OffsetUiListObserver

Represents an UiNodeListObserver that applies an offset to all indexes.

PanelList

Represents the final UI list in a panel layout node.

PanelListDataChanges

Identifies changes in PanelListData since last layout.

PanelListRange

First and last child widget in a PanelList.

SORTING_LIST

Represents the contextual parent SortingList during an update.

SortingList

Represents a sorted view into an UiNode::is_list.

UiNode

Represents an UI tree node instance.

UiVec

Vec of boxed UI nodes.

WeakNode

Weak reference to a ArcNode.

WhenUiNodeBuilder

Builds a node that can be one of multiple options, selected by the first condition that is true, or a fallback default.

WidgetUiNode

Extra UiNode methods for nodes that define a widget instance context.

ZIndex

Position of a widget inside an UI node list render operation.

Z_INDEX

Access to widget z-index in a parent PanelList.

Enums

UiNodeMethod

Identifies the UiNode UI methods.

UiNodeOp

Represents a node operation in a match_node.

Traits

IntoUiNode

Represents a value that can become a UiNode instance.

PanelListData

Represents an item’s associated data in a PanelList.

UiNodeImpl

Represents an UiNode implementation.

UiNodeListObserver

Represents an UiNode::update_list observer that can be used to monitor widget insertion, removal and re-order.

WidgetUiNodeImpl

Represents an UiNodeImpl that defines a widget instance scope.

Functions

extend_widget

Create a widget node that wraps the widget with any number of other non-widget nodes and still delegates with_context to the widget.

match_node

Creates a node that is implemented as a closure that matches over UiNodeOp and delegates to another child node.

match_node_leaf

Creates a node that is implemented as a closure that matches over UiNodeOp and does not delegate to any child node.

match_widget

Creates a widget that is implemented as a closure that matches over UiNodeOp and delegates to another child widget.

with_new_context_init_id

Wraps child in a node that provides a unique ContextInitHandle, refreshed every (re)init.