zng_app::widget::node

Function extend_widget

pub fn extend_widget(
    widget: impl IntoUiNode,
    build_extension: impl FnOnce(UiNode) -> UiNode,
) -> UiNode

Create a widget node that wraps the widget with any number of other non-widget nodes and still delegates with_context to the widget.

Note that the with_context is called in the context of widget, not in the context of the build_extension nodes. Other node operations are delegated to the build_extension nodes, and they in turn must delegate to the input child node that is widget.