zng_app::widget::node

Function match_node_leaf

pub fn match_node_leaf(
    closure: impl FnMut(UiNodeOp<'_>) + Send + 'static,
) -> UiNode

Creates a node that is implemented as a closure that matches over UiNodeOp and does not delegate to any child node.