Struct zng_app::widget::node::MatchNodeChildren

pub struct MatchNodeChildren<L> { /* private fields */ }

Children node of match_node_list.

When the closure does not delegate to this list the delegation happens after the closure returns. The UiNodeList methods that flag as delegated are all the *_all methods and the methods that access mutable references to each child and the UiNodeList::with_node. You can use the children accessor to visit children without flagging as delegated.

Implementations

impl<L: UiNodeList> MatchNodeChildren<L>

pub fn delegated(&mut self)

Flags the current operation as delegated, stopping the default delegation after the closure ends.

Note that each *_all method and the methods that give mutable access to children already flags this.

pub fn has_delegated(&self) -> bool

If the current operation was already delegated to the children.

pub fn children(&mut self) -> &mut L

Reference the children.

Note that if you delegate using this reference you must call delegated.