Struct zng_app::widget::node::PanelList

pub struct PanelList<D = DefaultPanelListData>
where
    D: PanelListData,
{ /* private fields */ }

Represents the final UiNodeList in a panel layout node.

Panel widgets should wrap their children list on this type to support z-index sorting and to easily track associated item data.

By default the associated item data is a DefaultPanelListData that represents the offset of each item inside the panel, but it can be any type that implements PanelListData. The panel list default render implementation uses this data to position the children widgets. Note that you must commit_data changes to this data at the end of a layout pass.

Panel widgets can also mark the list using track_info_range to implement getter properties such as is_odd or is_even.

Implementations

impl PanelList<DefaultPanelListData>

pub fn new(list: impl UiNodeList) -> Self

New from list and default data.

impl<D> PanelList<D>
where D: PanelListData,

pub fn new_custom(list: impl UiNodeList) -> Self

New from list and custom data type.

pub fn track_info_range( self, info_id: impl Into<StateId<PanelListRange>> ) -> Self

Enable tracking the first and last child in the parent widget info.

The info is set in the info_id, it can be used to identify the children widgets that are the panel children as the info tree may track extra widgets as children when they are set by other properties, like background.

pub fn into_parts( self ) -> (BoxedUiNodeList, Vec<Mutex<D>>, FrameValueKey<PxTransform>, Option<StateId<PanelListRange>>)

Into list and associated data.

pub fn from_parts( list: BoxedUiNodeList, data: Vec<Mutex<D>>, offset_key: FrameValueKey<PxTransform>, info_id: Option<StateId<PanelListRange>> ) -> Self

New from list and associated data.

Panics

Panics if the list and data don’t have the same length.

pub fn info_id(&self) -> Option<StateId<PanelListRange>>

Gets the ID set on the parent widget info if track_info_range was enabled.

pub fn with_node<R, F>(&mut self, index: usize, f: F) -> R
where F: FnOnce(&mut BoxedUiNode, &mut D) -> R,

Visit the specific node with associated data, panic if index is out of bounds.

pub fn for_each<F>(&mut self, f: F)
where F: FnMut(usize, &mut BoxedUiNode, &mut D),

Calls f for each node in the list with the index and associated data.

pub fn par_each<F>(&mut self, f: F)
where F: Fn(usize, &mut BoxedUiNode, &mut D) + Send + Sync, D: Sync,

Calls f for each node in the list with the index and associated data in parallel.

pub fn par_fold_reduce<T, I, F, R>( &mut self, identity: I, fold: F, reduce: R ) -> T
where T: Send + 'static, I: Fn() -> T + Send + Sync, F: Fn(T, usize, &mut BoxedUiNode, &mut D) -> T + Send + Sync, R: Fn(T, T) -> T + Send + Sync,

Calls fold for each node in the list with the index and associated data in parallel.

This method behaves the same as UiNodeList::par_fold_reduce, with the added data.

pub fn measure_each<F, S>( &mut self, wm: &mut WidgetMeasure, measure: F, fold_size: S ) -> PxSize
where F: Fn(usize, &mut BoxedUiNode, &mut D, &mut WidgetMeasure) -> PxSize + Send + Sync, S: Fn(PxSize, PxSize) -> PxSize + Send + Sync,

Call measure for each node and combines the final size using fold_size.

The call to measure can be parallel if Parallel::LAYOUT is enabled, the inputs are the child index, node, data and the WidgetMeasure.

pub fn layout_each<F, S>( &mut self, wl: &mut WidgetLayout, layout: F, fold_size: S ) -> PxSize
where F: Fn(usize, &mut BoxedUiNode, &mut D, &mut WidgetLayout) -> PxSize + Send + Sync, S: Fn(PxSize, PxSize) -> PxSize + Send + Sync,

Call layout for each node and combines the final size using fold_size.

The call to layout can be parallel if Parallel::LAYOUT is enabled, the inputs are the child index, node, data and the WidgetLayout.

pub fn for_each_z_sorted( &mut self, f: impl FnMut(usize, &mut BoxedUiNode, &mut D) )

Iterate over the list in the Z order.

pub fn z_map(&mut self, index: usize) -> usize

Gets the index sorted in the list.

pub fn data(&mut self, index: usize) -> &mut D

Reference the associated data.

pub fn commit_data(&mut self) -> PanelListDataChanges

Calls commit for each child data, aggregate changes.

This must be called after the last update to the children data in a layout pass. Note that you can call commit directly in a for_each iteration if that iteration is the last in the layout pass.

pub fn offset_key(&self) -> FrameValueKey<PxTransform>

Key used to define reference frames for each item.

The default implementation of render_all uses this key and the item index.