Struct zng_app::widget::node::SortingList

pub struct SortingList { /* private fields */ }

Represents a sorted view into an UiNodeList.

The underlying list is not changed, a sorted index map is used to iterate the underlying list.

Note that the *_all methods are not sorted, only the other accessors map to the sorted position of nodes. The sorting is lazy and gets invalidated on every init and every time there are changes observed in update_all.

Implementations

impl SortingList

pub fn new( list: impl UiNodeList, sort: impl Fn(&mut BoxedUiNode, &mut BoxedUiNode) -> Ordering + Send + 'static, ) -> Self

New from list and sort function.

pub fn list(&mut self) -> &mut BoxedUiNodeList

Mutable borrow the inner list.

You must call invalidate_sort if any modification is done to the list.

pub fn invalidate_sort(&mut self)

Invalidate the sort, the list will resort on the nest time the sorted positions are needed.

Note that you can also invalidate sort from the inside using SORTING_LIST::invalidate_sort.