zng_app::widget

Trait ResponseVarSubscribe

pub trait ResponseVarSubscribe<T: VarValue> {
    // Required methods
    fn on_pre_rsp<H>(&self, handler: H) -> VarHandle
       where H: AppHandler<OnVarArgs<T>>;
    fn on_rsp<H>(&self, handler: H) -> VarHandle
       where H: AppHandler<OnVarArgs<T>>;
}

Extension methods to subscribe app handlers to a response variable.

Required Methods

fn on_pre_rsp<H>(&self, handler: H) -> VarHandle
where H: AppHandler<OnVarArgs<T>>,

Add a handler that is called once when the response is received, the handler is called before all other UI updates.

The handle is not called if already is_done, in this case a dummy handle is returned.

fn on_rsp<H>(&self, handler: H) -> VarHandle
where H: AppHandler<OnVarArgs<T>>,

Add a handler that is called once when the response is received, the handler is called after all other UI updates.

The handle is not called if already is_done, in this case a dummy handle is returned.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.