zng_app

Module crash_handler

App-process crash handler.

See the zng::app::crash_handler documentation for more details.

Macros

crash_handler_config

Register a FnOnce(&mut CrashConfig) closure to be called on process init to configure the crash handler.

Structs

BacktraceFrame

Represents a frame parsed from a stack backtrace.

CrashArgs

Arguments for the crash handler dialog function.

CrashConfig

Crash handler config.

CrashError

Info about an app-process crash.

CrashPanic

Panic parsed from a stderr dump.

Constants

NO_CRASH_HANDLER

Environment variable that causes the crash handler to not start if set.

Functions

remove_ansi_csi

Remove ANSI escape sequences (CSI) from s.

restart_count

Gets the number of crash restarts in the app-process.