zng_app

Module event

App event and commands API.

Re-exports

pub use command;

pub use command;

Macros

command

Declares new Command static items.

event

Declares new Event<A> static items.

event_args

Declares new EventArgs types.

Structs

AnyEvent

Represents an Event without the args type.

AppCommandArgs

Arguments for Command::on_event handler closure.

BuildFxHasher

Const rustc-hash hasher.

Command

Identifies a command event.

CommandArgs

Event args for command events.

CommandHandle

A handle to a Command subscription.

CommandMeta

Access to metadata of a command.

CommandMetaVarId

Unique identifier of a command metadata state variable.

CommandParam

Represents a reference counted dyn Any object parameter for a command request.

EVENTS

App events and commands service.

EVENTS_L10N

EVENTS L10N integration.

Event

Represents an event.

EventHandle

Represents an event widget subscription, handler callback or hook.

EventHandles

Represents a collection of event handles.

EventPropagationHandle

Event propagation handle associated with one or multiple EventArgs.

EventSender

An event update sender that can be used from any thread and without access to EVENTS.

Enums

CommandScope

Represents the scope of a Command.

Traits

AnyEventArgs

Methods of EventArgs that are object safe.

CommandInfoExt

Adds the info command metadata.

CommandNameExt

Adds the name command metadata.

EventArgs

Event<A> arguments.

Type Aliases

CommandMetaVar

Read-write command metadata variable.

CommandSet

Registered commands set.

ReadOnlyCommandMetaVar

Read-only command metadata variable.