zng_app

Macro property_args

macro_rules! property_args {
    ($($property:ident)::+ = $($value:tt)*) => { ... };
    ($($property:ident)::+ ::<$($generics:ty),*> = $($value:tt)*) => { ... };
    ($property:ident $(;)?) => { ... };
}

New Box<PropertyArgs> box from a property and value.

Syntax

The syntax is similar to a property assign in a widget.

property::path = <value>; - Args for the property function.

property::path; - Args for property with input of the same name, path here.

The <value> is the standard property init expression or named fields patterns that are used in widget assigns.

property = "value-0", "value-1"; - Unnamed args.

property = { value_0: "value-0", value_1: "value-1" } - Named args.

Panics

Panics if unset! is used as property value.