zng_app

Module timer

App timers, deadlines and timeouts.

The primary struct of this module is TIMERS. You can use it to create UI bound timers that run using only the main thread and can awake the app event loop to notify updates.

Structs

DeadlineArgs

Arguments for the handler of on_deadline.

DeadlineHandle

Represents a on_deadline handler.

TIMERS

App timers, deadlines and timeouts.

Timer

Represents a timer state in a TimerVar or interval handler.

TimerArgs

Arguments for an on_interval handler.

TimerHandle

Represents a on_interval handler.

WeakDeadlineHandle

Weak DeadlineHandle

WeakTimerHandle

Weak TimerHandle.

Type Aliases

DeadlineVar

A deadline timer.

TimerVar

An interval timer.