zng_app

Module update

App updates API.

Structs

ContextUpdates

Updates that must be reacted by an app owner.

EventUpdate

Represents a single event update.

InfoUpdates

Widget info updates of the current cycle.

LayoutUpdates

Widget layout updates of the current cycle.

OnUpdateHandle

Represents an on_pre_update or on_update handler.

RenderUpdates

Widget render updates of the current cycle.

UPDATES

Update schedule service.

UpdateArgs

Arguments for an on_pre_update, on_update or run handler.

UpdateDeliveryList

Represents all the widgets and windows marked to receive an update.

WeakOnUpdateHandle

Weak OnUpdateHandle.

WidgetUpdates

Widget updates of the current cycle.

Enums

UpdateOp

Identify node and app-extension operations that can be requested.

Traits

UpdateSubscribers

Represents a set of widgets that subscribe to an event source.

UpdatesTraceUiNodeExt

Extension methods for infinite loop diagnostics.

WidgetPathProvider

Provides an iterator of widget IDs and a window ID.

Functions

updates_trace_event

Custom log entry in the app loop diagnostics.

updates_trace_span

Custom span in the app loop diagnostics.