zng_app

Module view_process

View process connection and other types.

Modules

raw_device_events

Events directly from view-process not targeting any windows.

raw_events

Events directly from the view-process targeting the app windows.

Structs

VIEW_PROCESS

Connection to the running view-process for the context app.

ViewClipboard

View-process clipboard methods.

ViewHeadless

Handle to a headless surface/document open in the View Process.

ViewImage

Handle to an image loading or loaded in the View Process.

ViewProcessInitedArgs

Arguments for the VIEW_PROCESS_INITED_EVENT.

ViewProcessSuspendedArgs

Arguments for the VIEW_PROCESS_SUSPENDED_EVENT.

ViewRenderer

Weak handle to a window or view.

ViewWindow

Handle to a window open in the view-process.

WeakViewImage

Connection to an image loading or loaded in the View Process.

WindowOpenData

Information about a successfully opened window.

Enums

EncodeError

Error returned by ViewImage::encode.

ViewWindowOrHeadless

View window or headless surface.

Statics

VIEW_PROCESS_INITED_EVENT

View-Process finished initializing and is now connected and ready.

VIEW_PROCESS_SUSPENDED_EVENT

View-Process suspended, all resources dropped.