zng_app::view_process

Module raw_events

Events directly from the view-process targeting the app windows.

These events get processed by app extensions to generate the events used in widgets, for example the KeyboardManager uses the RAW_KEY_INPUT_EVENT into focus targeted events.

Synthetic Input

You can notify these events to fake hardware input, please be careful that you mimic the exact sequence a real hardware would generate, app extensions can assume that the raw events are correct. The InputDeviceId for fake input must be unique but constant for each distinctive synthetic event source.

Structs

LowMemoryArgs

Arguments for LOW_MEMORY_EVENT.

RawAnimationsConfigChangedArgs

Arguments for the RAW_ANIMATIONS_CONFIG_CHANGED_EVENT.

RawAppDragEndedArgs

Arguments for the RAW_APP_DRAG_ENDED_EVENT.

RawAxisMotionArgs

Arguments for the RAW_AXIS_MOTION_EVENT.

RawChromeConfigChangedArgs

Arguments for the RAW_CHROME_CONFIG_CHANGED_EVENT.

RawColorsConfigChangedArgs

Arguments for the RAW_COLORS_CONFIG_CHANGED_EVENT.

RawDragCancelledArgs

Arguments for the RAW_DRAG_CANCELLED_EVENT.

RawDragDroppedArgs

Arguments for the RAW_DRAG_DROPPED_EVENT.

RawDragHoveredArgs

Arguments for the RAW_DRAG_HOVERED_EVENT.

RawDragMovedArgs

Arguments for the RAW_DRAG_MOVED_EVENT.

RawExtensionEventArgs

Arguments for the RAW_EXTENSION_EVENT.

RawFontAaChangedArgs

Arguments for the RAW_FONT_AA_CHANGED_EVENT.

RawFontChangedArgs

RAW_FONT_CHANGED_EVENT arguments.

RawFrameImageReadyArgs

Arguments for the RAW_FRAME_IMAGE_READY_EVENT.

RawFrameRenderedArgs

Arguments for the RAW_FRAME_RENDERED_EVENT.

RawHeadlessOpenArgs

Arguments for the RAW_HEADLESS_OPEN_EVENT.

RawImageArgs

Arguments for the image events.

RawImeArgs

Arguments for the RAW_IME_EVENT.

RawKeyInputArgs

Arguments for the RAW_KEY_INPUT_EVENT.

RawKeyRepeatConfigChangedArgs

Arguments for the RAW_KEY_REPEAT_CONFIG_CHANGED_EVENT.

RawLocaleChangedArgs

Arguments for the RAW_LOCALE_CONFIG_CHANGED_EVENT.

RawMonitorsChangedArgs

Arguments for the RAW_MONITORS_CHANGED_EVENT.

RawMouseArgs

Arguments for the RAW_MOUSE_ENTERED_EVENT and RAW_MOUSE_LEFT_EVENT.

RawMouseInputArgs

Arguments for the RAW_MOUSE_INPUT_EVENT.

RawMouseMovedArgs

Arguments for the RAW_MOUSE_MOVED_EVENT.

RawMouseWheelArgs

Arguments for the RAW_MOUSE_WHEEL_EVENT.

RawMultiClickConfigChangedArgs

Arguments for the RAW_MULTI_CLICK_CONFIG_CHANGED_EVENT.

RawScaleFactorChangedArgs

Arguments for the RAW_SCALE_FACTOR_CHANGED_EVENT.

RawTouchArgs

Arguments for the RAW_TOUCH_EVENT.

RawTouchConfigChangedArgs

Arguments for the RAW_TOUCH_CONFIG_CHANGED_EVENT.

RawTouchpadPressureArgs

Arguments for the RAW_TOUCHPAD_PRESSURE_EVENT.

RawWindowChangedArgs

Arguments for the RAW_WINDOW_CHANGED_EVENT.

RawWindowCloseArgs

Arguments for the RAW_WINDOW_CLOSE_EVENT.

RawWindowCloseRequestedArgs

Arguments for the RAW_WINDOW_CLOSE_REQUESTED_EVENT.

RawWindowFocusArgs

Arguments for the RAW_WINDOW_FOCUS_EVENT.

RawWindowOpenArgs

Arguments for the RAW_WINDOW_OPEN_EVENT.

RawWindowOrHeadlessOpenErrorArgs

Arguments for the RAW_WINDOW_OR_HEADLESS_OPEN_ERROR_EVENT.

Statics

LOW_MEMORY_EVENT

System low memory warning, some platforms may kill the app if it does not release memory.

RAW_ANIMATIONS_CONFIG_CHANGED_EVENT

Change in system animation enabled config.

RAW_APP_DRAG_ENDED_EVENT

Drag & drop operation started by the app has dropped or was cancelled.

RAW_AXIS_MOTION_EVENT

Motion on some analog axis send to a window.

RAW_CHROME_CONFIG_CHANGED_EVENT

System window chrome config changed.

RAW_COLORS_CONFIG_CHANGED_EVENT

Color scheme or accent color preference changed for a window.

RAW_DRAG_CANCELLED_EVENT

Data was dragged away from the window or the operation was cancelled.

RAW_DRAG_DROPPED_EVENT

Data was drag-dropped on a window.

RAW_DRAG_HOVERED_EVENT

Data was dragged over a window.

RAW_DRAG_MOVED_EVENT

Data dragging over the window has moved.

RAW_EXTENSION_EVENT

Custom view-process extension event.

RAW_FONT_AA_CHANGED_EVENT

Change in system font anti-aliasing config.

RAW_FONT_CHANGED_EVENT

Change in system text fonts, install or uninstall.

RAW_FRAME_IMAGE_READY_EVENT

Image generated from a frame is ready for reading.

RAW_FRAME_RENDERED_EVENT

A frame finished rendering and was presented in a window.

RAW_HEADLESS_OPEN_EVENT

A headless surface has finished initializing in the view-process.

RAW_IMAGE_LOADED_EVENT

Image loaded without errors.

RAW_IMAGE_LOAD_ERROR_EVENT

Image failed to load.

RAW_IMAGE_METADATA_LOADED_EVENT

Image metadata loaded without errors.

RAW_IMAGE_PARTIALLY_LOADED_EVENT

Progressively decoded image has decoded more pixels.

RAW_IME_EVENT

An IME event was received by a window.

RAW_KEY_INPUT_EVENT

A key press or release targeting a window.

RAW_KEY_REPEAT_CONFIG_CHANGED_EVENT

Change in system key repeat interval config.

RAW_LOCALE_CONFIG_CHANGED_EVENT

Change in system locale config.

RAW_MONITORS_CHANGED_EVENT

Monitors added, removed or modified.

RAW_MOUSE_ENTERED_EVENT

Mouse pointer started hovering a window.

RAW_MOUSE_INPUT_EVENT

Mouse button was pressed or released when the mouse was over a window.

RAW_MOUSE_LEFT_EVENT

Mouse pointer stopped hovering a window.

RAW_MOUSE_MOVED_EVENT

Mouse pointer moved over a window.

RAW_MOUSE_WHEEL_EVENT

Mouse wheel scrolled when the mouse was over a window.

RAW_MULTI_CLICK_CONFIG_CHANGED_EVENT

Change in system “double-click” config.

RAW_SCALE_FACTOR_CHANGED_EVENT

Pixel scale factor for a monitor screen and its windows has changed.

RAW_TOUCHPAD_PRESSURE_EVENT

Touchpad touched when the mouse was over a window.

RAW_TOUCH_CONFIG_CHANGED_EVENT

Change in system touch config.

RAW_TOUCH_EVENT

A window was touched.

RAW_WINDOW_CHANGED_EVENT

A window was moved, resized or has a state change.

RAW_WINDOW_CLOSE_EVENT

A window was destroyed.

RAW_WINDOW_CLOSE_REQUESTED_EVENT

A window was requested to close.

RAW_WINDOW_FOCUS_EVENT

A window received or lost focus.

RAW_WINDOW_OPEN_EVENT

A window has finished initializing in the view-process.

RAW_WINDOW_OR_HEADLESS_OPEN_ERROR_EVENT

A window or headless surface initialization failed in the view-process.