Module zng_app::widget

Widget, UI node API.

Modules

base

The widget base, nodes and properties used in most widgets.

border

Border and line types.

builder

Widget and property builder types.

info

Widget info tree.

inspector

Helper types for inspecting an UI tree.

node

Widget nodes types, UiNode, UiNodeList and others.

Macros

widget_impl

Implement a property on the widget to strongly associate it with the widget.

widget_set

Sets properties and when condition on a widget builder.

Structs

OnVarArgs

Arguments for a var event handler.

StaticWidgetId

Lazy inited WidgetId.

WIDGET

Current context widget.

WidgetCtx

Defines the backing data of WIDGET.

WidgetHandlesCtx

Defines the backing data for WIDGET.with_handles.

WidgetId

Unique ID of a widget.

Enums

WidgetUpdateMode

Defines how widget update requests inside WIDGET::with_context are handled.

Traits

AnyVarSubscribe

Extension method to subscribe any widget to a variable.

ResponseVarSubscribe

Extension methods to subscribe app handlers to a response variable.

UiTaskWidget

Integrate UiTask with widget updates.

VarLayout

Extension methods to layout var values.

VarSubscribe

Extension methods to subscribe any widget to a variable or app handlers to a variable.

Attribute Macros

easing

Expands a property assign to include a build action that applies an easing transition to the variable inputs.

property

Expands a function to a widget property.

ui_node

Expands an impl into a UiNode impl.

widget

Expands a struct to a widget and macro.

widget_mixin

Expands a generic struct to a widget mixin.