Function zng_app::widget::base::id

pub fn id(id: impl IntoValue<WidgetId>)

c Unique ID of the widget instance.

Note that the id can convert from a &'static str unique name.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.