Module zng_app::widget::border

Border and line types.

Structs

BORDER

Coordinates nested borders and corner-radius.

BorderSide

The line style and color for the sides of a widget’s border.

BorderSides

The line style and color for each side of a widget’s border.

CornerRadius

Radius of each corner of a border defined from Size values.

Enums

BorderStyle

The line style for the sides of a widget’s border.

CornerRadiusFit

Defines how the corner radius is computed for each usage.

LineOrientation

Orientation of a straight line.

LineStyle

Represents a line style.

Statics

BORDER_ALIGN_VAR

How much a widget’s border offsets affects the widget’s fill content.

BORDER_OVER_VAR

If the border is rendered over the child nodes.

CORNER_RADIUS_FIT_VAR

Corner radius fit.

CORNER_RADIUS_VAR

Corner radius.