zng_app::widget::info
Module access

Accessibility metadata types.

Structs

AccessBuildArgs

Arguments for on_access_build handlers.

AccessEnabled

Defines how accessibility info is enabled.

AutoComplete

Defines how inputting text could trigger display of one or more predictions of the user’s intended value.

Invalid

Defines the kind of invalid data error of a widget.

WidgetAccessInfo

Accessibility info for a widget.

WidgetAccessInfoBuilder

Accessibility metadata.

Enums

AccessCmdName

Accessibility command without associated data.

AccessRole

Accessibility role of a node in the accessibility tree.

CurrentKind

Kind of current item a widget represents.

LiveIndicator

Defines how a live update is communicated to the user.

Orientation

Widget orientation.

Popup

Popup type.

SortDirection

Sort direction.