Module zng_app::widget::inspector

Helper types for inspecting an UI tree.

When compiled with the "inspector" feature all widget instances are instrumented with inspection node that shares a clone of the WidgetBuilder in the WidgetInfo.

Structs

InspectorActualVars

Inspected contextual variables actualized at the moment of info build.

InspectorInfo

Widget instance inspector info.

Enums

InstanceItem

Widget instance item.

Traits

InspectPropertyPattern

Query pattern for the WidgetInfoInspectorExt inspect methods.

InspectWidgetPattern

Query pattern for the WidgetInfoInspectorExt inspect methods.

WidgetInfoInspectorExt

Extensions methods for WidgetInfo.