Module zng_app::widget::node

Widget nodes types, UiNode, UiNodeList and others.

Macros

ui_vec

Creates an UiNodeVec containing the arguments.

Structs

AdoptiveChildNode

Placeholder for the dynamic child of an adoptive node.

AdoptiveNode

Represents a node setup to dynamically swap child.

ArcNode

A reference counted UiNode.

ArcNodeList

A reference counted UiNodeList.

DefaultPanelListData

Default PanelList associated data.

EditableUiNodeList

Represents an UiNodeList that can be modified using a connected sender.

EditableUiNodeListRef

Represents a sender to an EditableUiNodeList.

FillUiNode

A UI node that fills the available layout space.

MatchNodeChild

Child node of match_node.

MatchNodeChildren

Children node of match_node_list.

MatchWidgetChild

Child node of match_widget.

NilUiNode

A UI node that does nothing and has collapsed layout (zero size).

OffsetUiListObserver

Represents an UiNodeListObserver that applies an offset to all indexes.

PanelList

Represents the final UiNodeList in a panel layout node.

PanelListDataChanges

Identifies changes in PanelListData since last layout.

PanelListRange

First and last child widget in a PanelList.

SORTING_LIST

Represents the contextual parent SortingList during an update.

SortingList

Represents a sorted view into an UiNodeList.

UiNodeListChainImpl

Implementation of UiNodeListChain::chain.

UiNodeVec

Vec of boxed UI nodes.

WeakNode

Weak reference to a ArcNode<U>.

WeakNodeList

Weak reference to a ArcNodeList<U>.

WhenUiNodeBuilder

Builds a node that can be one of multiple options, selected by the first condition that is true, or a fallback default.

WhenUiNodeListBuilder

Builds a node list that can be one of multiple options, selected by the first condition that is true, or a fallback default.

ZIndex

Position of a widget inside an UiNodeList render operation.

Z_INDEX

Access to widget z-index in a parent PanelList.

Enums

UiNodeOp

Represents a node operation in a match_node.

UiNodeOpMethod

Identifies an UiNodeOp method without the associated data.

Traits

PanelListData

Represents an item’s associated data in a PanelList.

UiNode

Represents an UI tree node.

UiNodeList

Represents a list of UI nodes.

UiNodeListChain

Adds the chain method for all UiNodeList implementors.

UiNodeListObserver

Represents an UiNodeList::update_all observer that can be used to monitor widget insertion, removal and re-order.

Functions

extend_widget

Create a widget node that wraps the widget with any number of other non-widget nodes and still delegates with_context to the widget.

match_node

Creates a node that is implemented as a closure that matches over UiNodeOp and delegates to another child node.

match_node_leaf

Creates a node that is implemented as a closure that matches over UiNodeOp and does not delegate to any child node.

match_node_list

Creates a node that is implemented as a closure that matches over UiNodeOp and delegates to multiple children nodes in a list.

match_node_typed

Like match_node, but does not change the child type when build with dyn_node.

match_widget

Creates a widget that is implemented as a closure that matches over UiNodeOp and delegates to another child widget.

with_new_context_init_id

Wraps child in a node that provides a unique ContextInitHandle, refreshed every (re)init.

Type Aliases

BoxedUiNode

An UiNode in a box.

BoxedUiNodeList

An UiNodeList in a box.