Struct zng_app::widget::node::MatchNodeChild

pub struct MatchNodeChild<C> { /* private fields */ }

Child node of match_node.

When the closure does not delegate to this node the delegation happens after the closure returns.

Implementations

impl<C: UiNode> MatchNodeChild<C>

pub fn delegated(&mut self)

Flags the current operation as delegated, stopping the default delegation after the closure ends.

Note that each node operation methods already flags this.

pub fn has_delegated(&self) -> bool

If the current operation was already delegated to the child.

pub fn child(&mut self) -> &mut C

Borrow the actual child.

Note that if you delegate using this reference you must call delegated.

Warning: that match_node changes the child type to BoxedUiNode when build with the "dyn_node" feature. To get a consistent child type use the BoxedUiNode directly or use match_node_typed.