zng_app::widget::node

Struct PanelList

pub struct PanelList<D = DefaultPanelListData>
where
    D: PanelListData,
{ /* private fields */ }

Represents the final UI list in a panel layout node.

Panel widgets should wrap their children node on this type to support Z-index sorting and to easily track associated item data.

By default the associated item data is a DefaultPanelListData that represents the offset of each item inside the panel, but it can be any type that implements PanelListData. The panel list default render implementation uses this data to position the children widgets. Note that you must commit_data changes to this data at the end of a layout pass.

Panel widgets can also mark the list using track_info_range to implement getter properties such as is_odd or is_even.

Implementations

impl PanelList<DefaultPanelListData>

pub fn new(list: impl IntoUiNode) -> Self

New from list and default data.

impl<D> PanelList<D>
where D: PanelListData,

pub fn new_custom(list: impl IntoUiNode) -> Self

New from list and custom data type.

pub fn track_info_range( self, info_id: impl Into<StateId<PanelListRange>>, ) -> Self

Enable tracking the first and last child in the parent widget info.

The info is set in the info_id, it can be used to identify the children widgets that are the panel children as the info tree may track extra widgets as children when they are set by other properties, like background.

pub fn into_parts( self, ) -> (UiNode, Vec<Mutex<D>>, FrameValueKey<PxTransform>, Option<StateId<PanelListRange>>)

Into list and associated data.

pub fn from_parts( list: UiNode, data: Vec<Mutex<D>>, offset_key: FrameValueKey<PxTransform>, info_id: Option<StateId<PanelListRange>>, ) -> Self

New from list and associated data.

Panics

Panics if the list and data don’t have the same length.

pub fn info_id(&self) -> Option<StateId<PanelListRange>>

Gets the ID set on the parent widget info if track_info_range was enabled.

pub fn with_child<R>( &mut self, index: usize, visitor: impl FnOnce(&mut UiNode, &mut D) -> R, ) -> R

Call visitor with a exclusive reference to the child node and associated data identified by index.

Panics if the index is out of bounds.

pub fn for_each_child( &mut self, visitor: impl FnMut(usize, &mut UiNode, &mut D), )

Call visitor for each child node of self, one at a time.

The closure parameters are the child index, the child and the associated data.

pub fn try_for_each_child<B: VarValue>( &mut self, visitor: &mut dyn FnMut(usize, &mut UiNode, &mut D) -> ControlFlow<B>, ) -> ControlFlow<B>

Call visitor for each child node of self, one at a time, with control flow.

The closure parameters are the child index, the child and the associated data.

pub fn par_each_child( &mut self, visitor: impl Fn(usize, &mut UiNode, &mut D) + Sync, )
where D: Sync,

Calls visitor for each child node in parallel.

The closure parameters are the child index, the child and the associated data.

pub fn par_fold_reduce<T>( &mut self, identity: T, fold: impl Fn(T, usize, &mut UiNode, &mut D) -> T + Sync, reduce: impl Fn(T, T) -> T + Send + Sync, ) -> T
where T: VarValue,

Calls fold for each child node with associated data in parallel, with fold accumulators produced by cloning identity, then merges the folded results using reduce to produce the final value also in parallel.

The reduce call is associative, the order is preserved in the result.

pub fn measure_list( &mut self, wm: &mut WidgetMeasure, measure: impl Fn(usize, &mut UiNode, &mut D, &mut WidgetMeasure) -> PxSize + Sync, fold_size: impl Fn(PxSize, PxSize) -> PxSize + Sync, ) -> PxSize

Call measure for each node with associated data and combines the final size using fold_size.

pub fn layout_list( &mut self, wl: &mut WidgetLayout, layout: impl Fn(usize, &mut UiNode, &mut D, &mut WidgetLayout) -> PxSize + Sync, fold_size: impl Fn(PxSize, PxSize) -> PxSize + Sync, ) -> PxSize

Call layout for each node with associated data and combines the final size using fold_size.

pub fn render_list( &mut self, frame: &mut FrameBuilder, render: impl Fn(usize, &mut UiNode, &mut D, &mut FrameBuilder) + Sync, )

Call render for each node with associated data.

Note that the PanelListData child offset and reference frame are already pushed when render is called.

pub fn render_update_list( &mut self, update: &mut FrameUpdate, render_update: impl Fn(usize, &mut UiNode, &mut D, &mut FrameUpdate) + Sync, )

Call render_update for each node with associated data.

Note that the PanelListData child offset and reference frame are already pushed when render_update is called.

pub fn for_each_z_sorted( &mut self, visitor: impl FnMut(usize, &mut UiNode, &mut D), )

Iterate over the list in the Z order.

pub fn z_map(&mut self, index: usize) -> usize

Gets the index sorted in the list.

pub fn data(&mut self, index: usize) -> &mut D

Reference the associated data.

pub fn commit_data(&mut self) -> PanelListDataChanges

Calls commit for each child data, aggregate changes.

This must be called after the last update to the children data in a layout pass. Note that you can call commit directly in a for_each iteration if that iteration is the last in the layout pass.

pub fn offset_key(&self) -> FrameValueKey<PxTransform>

Key used to define reference frames for each item.

The default implementation of render_all uses this key and the item index.