Trait zng_app::widget::node::PanelListData

pub trait PanelListData: Default + Send + Any {
    // Required methods
    fn child_offset(&self) -> PxVector;
    fn define_reference_frame(&self) -> bool;
    fn commit(&mut self) -> PanelListDataChanges;
}

Represents an item’s associated data in a PanelList.

Required Methods

fn child_offset(&self) -> PxVector

Gets the child offset to be used in the default render_all and render_update_all implementations.

fn define_reference_frame(&self) -> bool

If a new reference frame should be created for the item during render.

fn commit(&mut self) -> PanelListDataChanges

Commit child_offset and define_reference_frame changes.

Returns flags that indicate what values changed.

Object Safety

This trait is not object safe.