zng_app::widget::node

Trait WidgetUiNodeImpl

pub trait WidgetUiNodeImpl: UiNodeImpl {
    // Required method
    fn with_context(
        &mut self,
        update_mode: WidgetUpdateMode,
        visitor: &mut dyn FnMut(),
    );
}

Represents an UiNodeImpl that defines a widget instance scope.

Widget defining nodes implement this trait and UiNodeImpl::as_widget.

Required Methods

fn with_context( &mut self, update_mode: WidgetUpdateMode, visitor: &mut dyn FnMut(), )

Calls visitor with the WIDGET context of the widget instance defined by the node.

If update_mode is WidgetUpdateMode::Bubble the update flags requested for the widget in visitor will be copied to the caller widget context, otherwise they are ignored.